+++ lib/field/field.h
@@ -61,17 +61,37 @@ class field {
         void setInterpolationSlices();
 
     public:
+        /** The field data is stored in this Blitz array */
         blitz::Array<real, 3> F;
 
+        /** This string is used to identify the field, and is useful in file-writing */
         std::string fieldName;
 
+        /** xStag, yStag and zStag are the boolean flags which specify whether the field variable is at cell-center, face-center, edge-center and so on.
+         *  When the flag is true, the variable is shifted by half the grid-spacing.
+         *  This means that when all the flags are true, the variable is at the cell-center.
+         *  And when all the flags are false, the variable is placed at the vertex centers (collocated arrangement). */
+        //@{
         const bool xStag, yStag, zStag;
+        //@}
 
-        blitz::RectDomain<3> fCore, fBulk;
+        /** The core slice is a view of the field data excluding the ghost points surrounding it */
+        blitz::RectDomain<3> fCore;
+
+        /** The bulk slice is a view of the field data excluding the walls surrounding it.
+         *  This view is therefore a subset of the core slice. */
+        blitz::RectDomain<3> fBulk;
+
+        /** fCLft, fCRgt, fCFrt, fCBak, fCBot, fCTop are views of the field data generated by shifting the core slice
+         *  in all the three independent directions. These views are useful in calculating derivatives, like in divergence calculations. */
+        //@{
         blitz::RectDomain<3> fCLft, fCRgt;
         blitz::RectDomain<3> fCFrt, fCBak;
         blitz::RectDomain<3> fCBot, fCTop;
+        //@}
 
+        /** The wall slices are views of the field data showing only the wall points.
+         *  Thus, the core slice is a union of both bulk and wall slices. */
         blitz::Array<blitz::RectDomain<3>, 1> fWalls;
 
         blitz::Array<blitz::RectDomain<3>, 1> PcIntSlices, QvIntSlices;

[…]
