### 2020-02-08 9b93fec
Updated DOI field in the paper bibliography
