+++ resources/paper.bib
@@ -4,7 +4,7 @@
     journal={Proc. Camb. Phil. Soc.},
     year={1947},
     month={Jan},
-    doi={https://doi.org/10.1017/S0305004100023197},
+    doi={10.1017/S0305004100023197},
     volume={43},
     number={1},
     pages={50--67},
@@ -17,7 +17,7 @@
     volume = {8},
     number = {12},
     pages = {2182-2189},
-    doi={https://doi.org/10.1063/1.1761178},
+    doi={10.1063/1.1761178},
     year = {1965}
 }
 
@@ -27,7 +27,7 @@
     Month = {mar},
     Pages = {77--91},
     Title = {{Scaling of a Fast Fourier Transform and a pseudo-spectral fluid solver up to 196608 cores}},
-    doi={https://doi.org/10.1016/j.jpdc.2017.10.014},
+    doi={10.1016/j.jpdc.2017.10.014},
     Volume = {113},
     Year = {2018}
 }
@@ -40,7 +40,7 @@
 	number = {895},
 	pages = {499-521},
 	year = {1937},
-	doi = {https://doi.org/10.1098/rspa.1937.0036}
+	doi = {10.1098/rspa.1937.0036}
 }
 
 @book{Wesseling:MG2004,
@@ -49,7 +49,7 @@
     series={An Introduction to Multigrid Methods},
     year={2004},
     isbn={9781930217089},
-    doi={https://doi.org/10.1063/1.1761178},
+    doi={10.1063/1.1761178},
     publisher={R.T. Edwards}
 }
 
@@ -58,7 +58,7 @@
     author={Briggs, W.L. and Henson, V.E. and McCormick, S.F.},
     series={Other Titles in Applied Mathematics},
     year={2000},
-    doi={https://doi.org/10.1137/1.9780898719505},
+    doi={10.1137/1.9780898719505},
     publisher={Society for Industrial and Applied Mathematics}
 }
 
@@ -70,7 +70,7 @@
     number={3},
     pages={387--411},
     year={1982},
-    doi={https://doi.org/10.1016/0021-9991(82)90058-4},
+    doi={10.1016/0021-9991(82)90058-4},
     publisher={Academic Press}
 }
 
@@ -81,7 +81,7 @@
     number = {10},
     pages = {1787 - 1806},
     year = {1972},
-    doi={https://doi.org/10.1016/0017-9310(72)90054-3},
+    doi={10.1016/0017-9310(72)90054-3},
     author = {S.V Patankar and D.B Spalding}
 }
 
@@ -91,7 +91,7 @@
 	Pages = {16--18},
 	Title = {{Dissipation of Energy in Locally Isotropic Turbulence}},
 	Volume = {32},
-    doi={https://doi.org/10.1098/rspa.1991.0076},
+    doi={10.1098/rspa.1991.0076},
 	Year = {1941}
 }
 
@@ -103,7 +103,7 @@
     number={1},
     pages={32--74},
     year={1928},
-    doi={https://doi.org/10.1007/BF01448839},
+    doi={10.1007/BF01448839},
     publisher={Springer}
 }
 
@@ -111,7 +111,7 @@
 	Address = {Berlin Heidelberg},
 	Author = {Canuto, Claudio and Hussaini, M Yousuff and Quarteroni, Alfio and Zang, Thomas A},
 	Publisher = {Springer-Verlag},
-    doi={https://doi.org/10.1007/978-3-642-84108-8},
+    doi={10.1007/978-3-642-84108-8},
 	Title = {{Spectral Methods in Fluid Dynamics}},
 	Year = {1988},
 }
@@ -121,7 +121,7 @@
 	Author = {Ferziger, Joel H and Peric, Milovan},
 	Edition = {3},
 	Publisher = {Springer-Verlag},
[…]
