### 2020-02-08 #2084 @kyleniemeyer
Hi @roshansamuel, any suggestions for reviewers from our list (mentioned at the top) would be appreciated; otherwise, I will look for some shortly.

Also, could you edit your article to remove the leading `https://doi.org/` from the DOI entries for each reference?
