### 2020-02-10 #2084 @roshansamuel
Hi, @kyleniemeyer , the leading ``https://doi.org/`` have been removed from all the DOI entries. The missing DOI has been added as well.

From the list of reviewers, harpolea, mancellin, and dlagrava seem to have good overlap with the target audience of the solver. Hence I would like to suggest them as potential reviewers.
