### 2020-02-12 #2095 @whedon
[REVIEW]: SARAS: A general-purpose PDE solver for fluid dynamics
**Submitting author:** @roshansamuel (<a href="http://orcid.org/0000-0002-1280-9881">Roshan Samuel</a>)
**Repository:** <a href="https://github.com/roshansamuel/saras" target ="_blank">https://github.com/roshansamuel/saras</a>
**Version:** v1.0
