### 2020-02-17 #2095 @dlagrava
@roshansamuel I have been trying to compile the code on Ubuntu 19.04. It seems that the yaml-cpp-dev that I get is newer than the version you used. I get the following:
error: ‘class YAML::Parser’ has no member named ‘GetNextDocument’; did you mean ‘HandleNextDocument’

It looks like you need yaml-cpp 0.3 and I get the 0.5 version from the package manager. Could you correct to ask for that particular version? ie. do not say to use the debian one for instance, but use the one you propose to download.

Cheers!
