### 2020-02-18 #2095 @roshansamuel
@dlagrava Thank you for pointing this out. We had tested the code all along with yaml-cpp 0.3 and were not aware of this incompatibility with 0.5. We are working on updating the code (and the documentation), and we will get back to you soon.
