### 2020-02-21 28c666a
Switched to Ofast flag. Updated README with yaml compatibility issue. Minor updates to documentation
