### 2020-02-21 #2095 @roshansamuel
Hi @dlagrava we now indicate the yaml-cpp compatibility issues in the README. We are testing the new yaml API on a development branch and update the solver once it is ready.
