### 2020-02-24 #2095 @harpolea
@roshansamuel I've had a few issues installing the required libraries as well. I also had issues with the yaml version, and it took me a while to get the blitz library installed as it seems to require python 2 (fortunately I was able to modify a single print statement to get it to work with python 3). I also had to modify the Cmake file to tell it which HDF5 header file to include (it may have got confused because I have multiple versions installed on this machine? I'm not familiar with Cmake so not sure).

Now I've fixed those issues it seems to have compiled successfully, so shall start testing
