### 2020-02-25 #2095 @harpolea
@roshansamuel I've been trying to run the test problems and have been getting the following error. I've checked my yaml version and it's 0.3.0, so I'm not entirely sure what's gone wrong here?
```
[ 87%] Linking CXX executable ../../../../saras
[100%] Built target saras
terminate called after throwing an instance of 'YAML::InvalidScalar'
  what():  yaml-cpp: error at line 94, column 31: invalid scalar
[xrb:01852] *** Process received signal ***
[xrb:01852] Signal: Aborted (6)
[xrb:01852] Signal code:  (-6)
[xrb:01852] [ 0] /lib/x86_64-linux-gnu/libc.so.6(+0x3ef20)[0x7fc127860f20]
[xrb:01852] [ 1] /lib/x86_64-linux-gnu/libc.so.6(gsignal+0xc7)[0x7fc127860e97]
[xrb:01852] [ 2] /lib/x86_64-linux-gnu/libc.so.6(abort+0x141)[0x7fc127862801]
[xrb:01852] [ 3] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x9987e)[0x7fc12849987e]
[xrb:01852] [ 4] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0xa5486)[0x7fc1284a5486]
[xrb:01852] [ 5] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0xa54f1)[0x7fc1284a54f1]
[xrb:01852] [ 6] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0xa5745)[0x7fc1284a5745]
[xrb:01852] [ 7] ./saras(_ZN4YAMLrsIiEENS_9enable_ifINS_21is_scalar_convertibleIT_EEvE4typeERKNS_4NodeERS3_+0x88)[0x5585b2213978]
[xrb:01852] [ 8] ./saras(_ZN6parser9parseYAMLEv+0x153b)[0x5585b220dbfb]
[xrb:01852] [ 9] ./saras(_ZN6parserC1Ev+0xc6)[0x5585b2210206]
[xrb:01852] [10] ./saras(main+0x42)[0x5585b217a572]
[xrb:01852] [11] /lib/x86_64-linux-gnu/libc.so.6(__libc_start_main+0xe7)[0x7fc127843b97]
[xrb:01852] [12] ./saras(_start+0x2a)[0x5585b217c9ea]
[xrb:01852] *** End of error message ***
terminate called after throwing an instance of 'YAML::InvalidScalar'
  what():  yaml-cpp: error at line 94, column 31: invalid scalar
[xrb:01853] *** Process received signal ***
[xrb:01853] Signal: Aborted (6)
[xrb:01853] Signal code:  (-6)
[xrb:01853] [ 0] /lib/x86_64-linux-gnu/libc.so.6(+0x3ef20)[0x7f5b37e59f20]
[xrb:01853] [ 1] /lib/x86_64-linux-gnu/libc.so.6(gsignal+0xc7)[0x7f5b37e59e97]
[xrb:01853] [ 2] /lib/x86_64-linux-gnu/libc.so.6(abort+0x141)[0x7f5b37e5b801]
[xrb:01853] [ 3] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x9987e)[0x7f5b38a9287e]
[xrb:01853] [ 4] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0xa5486)[0x7f5b38a9e486]
[xrb:01853] [ 5] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0xa54f1)[0x7f5b38a9e4f1]
[xrb:01853] [ 6] [xrb:01851] *** Process received signal ***
terminate called after throwing an instance of 'YAML::InvalidScalar'
  what():  yaml-cpp: error at line 94, column 31: invalid scalar
/usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0xa5745)[0x7f5b38a9e745]
[xrb:01853] [ 7] [xrb:01851] Signal: Aborted (6)
[xrb:01851] Signal code:  (-6)
./saras(_ZN4YAMLrsIiEENS_9enable_ifINS_21is_scalar_convertibleIT_EEvE4typeERKNS_4NodeERS3_+0x88)[0x5599bdf62978]
[xrb:01853] [ 8] [xrb:01851] [ 0] ./saras(_ZN6parser9parseYAMLEv+0x153b)[0x5599bdf5cbfb]
[xrb:01853] [ 9] /lib/x86_64-linux-gnu/libc.so.6(+0x3ef20)[0x7f43d8c01f20]
[xrb:01851] [ 1] ./saras(_ZN6parserC1Ev+0xc6)[0x5599bdf5f206]
[xrb:01853] [10] /lib/x86_64-linux-gnu/libc.so.6(gsignal+0xc7)[0x./saras(main+0x42)[0x5599bdec9572]
[xrb:01853] [11] 7f43d8c01e97]
[xrb:01851] [ 2] /lib/x86_64-linux-gnu/libc.so.6(__libc_start_main+0xe7)[0x7f5b37e3cb97]
[xrb:01853] [12] /lib/x86_64-linux-gnu/libc.so.6(abort+0x141)[0x7f43d8c03801]
[xrb:01851] [ 3] ./saras(_start+0x2a)[0x5599bdecb9ea]
[xrb:01853] *** End of error message ***
/usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x9987e)[0x7f43d983a87e]
[xrb:01851] [ 4] terminate called after throwing an instance of 'YAML::InvalidScalar'
  what():  yaml-cpp: error at line 94, column 31: invalid scalar
[xrb:01854] *** Process received signal ***
[xrb:01854] Signal: Aborted (6)
[xrb:01854] Signal code:  (-6)
/usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0xa5486)[0x7f43d9846486]
[xrb:01851] [ 5] [xrb:01854] [ 0] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0xa54f1)[0x7f43d98464f1]
[xrb:01851] [ 6] /lib/x86_64-linux-gnu/libc.so.6(+0x3ef20)[0x7f9e6bca4f20]
[xrb:01854] [ 1] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0xa5745)[0x7f43d9846745]
[xrb:01851] [ 7] /lib/x86_64-linux-gnu/libc.so.6(gsignal+0xc7)[0x7f9e6bca4e97]
[xrb:01854] [ 2] ./saras(_ZN4YAMLrsIiEENS_9enable_ifINS_21is_scalar_convertibleIT_EEvE4typeERKNS_4NodeERS3_+0x88)[0x56524cec9978]
[xrb:01851] [ 8] /lib/x86_64-linux-gnu/libc.so.6(abort+0x141)[0x7f9e6bca6801]
[xrb:01854] [ 3] ./saras(_ZN6parser9parseYAMLEv+0x153b)[0x56524cec3bfb]
[xrb:01851] [ 9] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x9987e)[0x7f9e6c8dd87e]
[xrb:01854] ./saras(_ZN6parserC1Ev+0xc6)[0x56524cec6206]
[xrb:01851] [10] [ 4] ./saras(main+0x42)[0x56524ce30572]
[xrb:01851] [11] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0xa5486)[0x7f9e6c8e9486]
[xrb:01854] [ 5] /lib/x86_64-linux-gnu/libc.so.6(__libc_start_main+0xe7)[0x7f43d8be4b97]
[xrb:01851] [12] ./saras(_start+0x2a)[0x56524ce329ea]
[xrb:01851] *** End of error message ***
/usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0xa54f1)[0x7f9e6c8e94f1]
[xrb:01854] [ 6] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0xa5745)[0x7f9e6c8e9745]
[xrb:01854] [ 7] ./saras(_ZN4YAMLrsIiEENS_9enable_ifINS_21is_scalar_convertibleIT_EEvE4typeERKNS_4NodeERS3_+0x88)[0x55fdf942e978]
[xrb:01854] [ 8] ./saras(_ZN6parser9parseYAMLEv+0x153b)[0x55fdf9428bfb]
[xrb:01854] [ 9] ./saras(_ZN6parserC1Ev+0xc6)[0x55fdf942b206]
[xrb:01854] [10] ./saras(main+0x42)[0x55fdf9395572]
[xrb:01854] [11] /lib/x86_64-linux-gnu/libc.so.6(__libc_start_main+0xe7)[0x7f9e6bc87b97]
[xrb:01854] [12] ./saras(_start+0x2a)[0x55fdf93979ea]
[xrb:01854] *** End of error message ***
--------------------------------------------------------------------------
mpirun noticed that process rank 0 with PID 0 on node xrb exited on signal 6 (Aborted).
--------------------------------------------------------------------------
validate_ldc.py:80: YAMLLoadWarning: calling yaml.load() without Loader=... is deprecated, as the default Loader is unsafe. Please read https://msg.pyyaml.org/load for full details.
  yamlData = yl.load(yamlFile)
Could not open file output/Soln_0030.0000.h5
```
