### 2020-02-26 #2095 @dlagrava
@roshansamuel I encountered the same problems as @harpolea . It is a good idea to have a CMake, but you can probably improve on that, especially because people may want to have a local blitz install etc. I also ran into the same problem:
```
[100%]` Built target saras
terminate called after throwing an instance of 'YAML::InvalidScalar'
  what():  yaml-cpp: error at line 94, column 31: invalid scalarterminate called after throwing an instance of 'YAML::InvalidScalar'
  what():  yaml-cpp: error at line 94, column 31: invalid scalar
terminate called after throwing an instance of 'YAML::InvalidScalar'
  what():  yaml-cpp: error at line 94, column 31: invalid scalar
terminate called after throwing an instance of 'YAML::InvalidScalar'
  what():  yaml-cpp: error at line 94, column 31: invalid scalar

[reReBoot:17122] *** Process received signal ***
[reReBoot:17122] Signal: Aborted (6)
[reReBoot:17122] Signal code:  (-6)
[reReBoot:17123] *** Process received signal ***
[reReBoot:17123] Signal: Aborted (6)
[reReBoot:17123] Signal code:  (-6)
[reReBoot:17120] *** Process received signal ***
[reReBoot:17120] Signal: Aborted (6)
[reReBoot:17120] Signal code:  (-6)
[reReBoot:17120] [ 0] /lib/x86_64-linux-gnu/libc.so.6(+0x3ef20)[0x7f9a116b0f20]
[reReBoot:17120] [ 1] [reReBoot:17121] *** Process received signal ***
[reReBoot:17121] Signal: Aborted (6)
[reReBoot:17121] Signal code:  (-6)
[reReBoot:17121] [ 0] [reReBoot:17122] [ 0] /lib/x86_64-linux-gnu/libc.so.6(+0x3ef20)[0x7f862a86df20]
[reReBoot:17122] [reReBoot:17123] [ 0] /lib/x86_64-linux-gnu/libc.so.6(+0x3ef20)[0x7f8840f55f20]
[reReBoot:17123] [ 1] /lib/x86_64-linux-gnu/libc.so.6(gsignal+0xc7)[0x7f8840f55e97]
[reReBoot:17123] [ 2] /lib/x86_64-linux-gnu/libc.so.6(gsignal+0xc7)[0x7f9a116b0e97]
[reReBoot:17120] [ 2] /lib/x86_64-linux-gnu/libc.so.6(+0x3ef20)[0x7f77b534af20]
[reReBoot:17121] [ 1] /lib/x86_64-linux-gnu/libc.so.6(gsignal[ 1] /lib/x86_64-linux-gnu/libc.so.6(gsignal+0xc7)[0x7f862a86de97]
[reReBoot:17122] [ 2] +0xc7)[0x7f77b534ae97]
[reReBoot:17121] [ 2] /lib/x86_64-linux-gnu/libc.so.6(abort+0x141)[0x7f9a116b2801]
[reReBoot:17120] [ 3] /lib/x86_64-linux-gnu/libc.so.6(abort+0x141)[0x7f862a86f801]
[reReBoot:17122] [ 3] /lib/x86_64-linux-gnu/libc.so.6(abort+0x141)[0x7f8840f57801]
[reReBoot:17123] [ 3] /lib/x86_64-linux-gnu/libc.so.6(abort+0x141)[0x7f77b534c801]
[reReBoot:17121] [ 3] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x8c957)[0x7f9a122d4957]
[reReBoot:17120] [ 4] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x8c957)[0x7f862b491957]
[reReBoot:17122] [ 4] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x8c957)[0x7f8841b79957]
[reReBoot:17123] [ 4] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x8c957)[0x7f77b5f6e957]
[reReBoot:17121] [ 4] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x92ab6)[0x7f8841b7fab6]
[reReBoot:17123] [ 5] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x92ab6)[0x7f9a122daab6]
[reReBoot:17120] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x92ab6)[0x7f862b497ab6]
[reReBoot:17122] [ 5] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x92af1)[0x7f8841b7faf1]
[reReBoot:17123] [ 6] [ 5] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x92ab6)[0x7f77b5f74ab6]
[reReBoot:17121] [ 5] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x92af1)[0x7f862b497af1]
[reReBoot:17122] [ 6] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x92af1)[0x7f9a122daaf1]
[reReBoot:17120] [ 6] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x92af1)[0x7f77b5f74af1]
[reReBoot:17121] [ 6] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x92d24)[0x7f8841b7fd24]
[reReBoot:17123] [ 7] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x92d24)[0x7f9a122dad24]
[reReBoot:17120] [ 7] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x92d24)[0x7f862b497d24]
[reReBoot:17122] [ 7] ./saras(_ZN4YAMLrsIiEENS_9enable_ifINS_21is_scalar_convertibleIT_EEvE4typeERKNS_4NodeERS3_+0x88)[0x561b327c3498]
[reReBoot:17123] [ 8] ./saras(_ZN4YAMLrsIiEENS_9enable_ifINS_21is_scalar_convertibleIT_EEvE4typeERKNS_4NodeERS3_+0x88)[0x55af1a046498]
[reReBoot:17120] [ 8] ./saras(_ZN6parser9parseYAMLEv+0x1747)[0x561b327bda57]
[reReBoot:17123] [ 9] ./saras(_ZN4YAMLrsIiEENS_9enable_ifINS_21is_scalar_convertibleIT_EEvE4typeERKNS_4NodeERS3_+0x88)[0x560d6873a498]
[reReBoot:17122] [ 8] /usr/lib/x86_64-linux-gnu/libstdc++.so.6(+0x92d24)[0x7f77b5f74d24]
./saras(_ZN6parser9parseYAMLEv+0x1747)[0x55af1a040a57]
[reReBoot:17120] [ 9] ./saras(_ZN6parserC2Ev+0xc6)[0x55af1a042d26]
[reReBoot:17120] [10] [reReBoot:17121] [ 7] ./saras(_ZN4YAMLrsIiEENS_9enable_ifINS_21is_scalar_convertibleIT_EEvE4typeERKNS_4NodeERS3_+0x88)[0x5564c18f6498]
[reReBoot:17121] [ 8] ./saras(_ZN6parserC2Ev+0xc6)[0x561b327bfd26]
[reReBoot:17123] [10] ./saras(main+0x42)[0x561b3272bb62]
[reReBoot:17123] [11] ./saras(_ZN6parser9parseYAMLEv+0x1747)[0x560d68734a57]
[reReBoot:17122] [ 9] ./saras(_ZN6parserC2Ev+0xc6./saras(main+0x42)[0x55af19faeb62]
[reReBoot:17120] [11] /lib/x86_64-linux-gnu/libc.so.6(__libc_start_main+0xe7)[0x7f9a11693b97]
[reReBoot:17120] [12] )[0x560d68736d26]
[reReBoot:17122] [10] ./saras(main+0x42)[0x560d686a2b62]
./saras(_ZN6parser9parseYAMLEv+0x1747)[0x5564c18f0a57]
[reReBoot:17121] [ 9] /lib/x86_64-linux-gnu/libc.so.6(__libc_start_main+0xe7)[0x7f8840f38b97]
[reReBoot:17123] [12] ./saras(_start+0x2a)[0x55af19faf6ea]
[reReBoot:17120] *** End of error message ***
./saras(_start+0x2a)[0x561b3272c6ea]
[reReBoot:17123] *** End of error message ***
[reReBoot:17122] [11] ./saras(_ZN6parserC2Ev+0xc6)[0x5564c18f2d26]
[reReBoot:17121] [10] /lib/x86_64-linux-gnu/libc.so.6(__libc_start_main+0xe7)[0x7f862a850b97]
[reReBoot:17122] [12] ./saras(_start+0x2a)./saras(main+0x42)[0x5564c185eb62]
[reReBoot:17121] [11] [0x560d686a36ea]
/lib/x86_64-linux-gnu/libc.so.6(__libc_start_main+0xe7)[0x7f77b532db97]
[reReBoot:17121] [12] [reReBoot:17122] *** End of error message ***
./saras(_start+0x2a)[0x5564c185f6ea]
[reReBoot:17121] *** End of error message ***
```
