### 2020-02-26 aaefad3
Updated the parameters file of the LDC testing module
