+++ lib/field/vfield.h
@@ -60,10 +60,8 @@ class vfield {
     public:
         field Vx, Vy, Vz;
 
-        /** derVx, derVy and derVz are three instances of the derivative class used to compute derivatives
-         *  When the flag is true, the variable is shifted by half the grid-spacing.
-         *  This means that when all the flags are true, the variable is at the cell-center.
-         *  And when all the flags are false, the variable is placed at the vertex centers (collocated arrangement). */
+        /** derVx, derVy and derVz are three instances of the derivative class used to compute derivatives.
+         *  They correspond to finite-differencing operations along X, Y and Z directions respectively. */
         //@{
         derivative derVx, derVy, derVz;
         //@}
