### 2020-02-26 #2095 @roshansamuel
Hi @harpolea and @dlagrava, thank you very much for bringing the Blitz installation issues to our notice. The latest version of Blitz available on GitHub at blitzpp/blitz has resolved the Python3 related issues. We are updating the installer provided by the link in documentation with this version of Blitz.

The YAML related error thrown when running the test module happened because I had failed to update the parameters YAML file of the LDC test input. I have corrected this and pushed an update to the repository. Many apologies for this lapse.

The error should be resolved now after a git pull.
