### 2020-03-01 #2095 @dlagrava
@roshansamuel the fix allows me to execute the code and I can run the tests. As a general rule, and I swear I remember a requirement like this on my latest review, we should test a release that you have made sure compiles and works. I will be going through the document and the rest of the review now. Thanks!
