### 2020-04-14 #2095 @arfon
@dlagrava - please, no apology is necessary. I totally understand that everyone’s situation is different right now, hence I’ve been checking in on all active JOSS reviews to see how authors, editors, and reviewers are getting on.

Please take your time and focus on your personal situation.
