### 2020-05-29 roshansamuel/saras#1 @dlagrava
JOSS 1: Community guidelines
Hello,

In order to be accepted as a JOSS submission, there need to be clear guidelines on how to contribute or report bugs to the repository. Could you please add this to the README?

Thanks,
Daniel
