### 2020-05-29 roshansamuel/saras#2 @dlagrava
JOSS 2: Documentation on how to write examples
Hello,

I have been reviewing your software. I understand that there are 2 tests that you can run. You can see some of the declarations to make a test there. However, I do not see any documentation on how to write my own example: it would be good to have an document that says how to setup a 2D/3D simulation, how to set the boundary conditions etc.

Cheers,
Daniel
