### 2020-05-29 roshansamuel/saras#3 @dlagrava
JOSS 3: Software intent and comparison to other software
Hello,

I have been reviewing your software. As part of the checks that I need to perform, I need to understand what this software has been written for ie. do you have a particular paper that needs it, because other software do not work? Why did you write the software?

Also, it would be good to explain and discuss why you chose to go with something like finite differences, which are out of favor on most popular CFD solvers.

Cheers,
Daniel
