### 2020-05-29 #2095 @dlagrava
@roshansamuel Sorry for the delay, it is unprecedented times that we are currently living, hope that you are well.

I have mostly finished my review. I have 3 outstanding issues that I have opened on your repository about community guidelines, some missing documentation to write examples and intent of the software and comparison with existing software. I am happy to continue the discussion on each of the issues individually.

Thanks!
