### 2020-06-02 0d5034e
Updated README with guidelines for contribution and bug reports
