### 2020-06-02 roshansamuel/saras#1 @roshansamuel
Hi,

Thank you for pointing this out. In commit 0d5034e to master, we have updated the README with guidelines for contributing and reporting bugs. Can this issue be marked as closed?

Thanks,

Roshan
