### 2020-06-04 #2095 @roshansamuel
Dear @kyleniemeyer, @dlagrava and @harpolea: In a recent discussion among the authors, we converged that two contributors, whose inputs were essential to the design of the solver, deserve to be included in the list of authors. Prof. Ravi Samtaney gave valuable ideas and greatly aided the development of multigrid pressure correction solver. At the same time, Prof. Fahad Anwer helped us in implementing the semi-implicit Crank-Nicholson scheme for time-integration.

We should have had included them in the beginning itself, but it was an oversight. As you all know, code development involves many personnel, and authorships tend to be confusing.  During the lockdown, we had a careful review of our work and the present paper; that’s when we came to the above decision. Kindly note that the code remains unchanged (except for some more documentation, as advised by the reviewers).

We request you to allow us to include the two original contributors to SARAS as authors to the present paper. Kindly let us know if you need any clarifications.
