### 2020-06-04 #2095 @kyleniemeyer
Hi @roshansamuel, that seems like a perfectly valid reason to me, as long as all the original authors agree to this. JOSS generally leaves authorship decisions up to the authors themselves, unless something seems off.
