### 2020-06-06 e657953
Updated the authors list in paper.md
