### 2020-06-26 147c174
Added section in README on setting up a new case, and documentation comments to the compilation script and YAML file
