+++ README.md
@@ -113,6 +113,65 @@ The following Python modules are necessary for the Python test script to execute
 
 At the end of the test, a plot of the x and y velocity profiles is shown to the user and saved as ``ldc_validation.png`` in the folder ``tests/ldcTest/``.
 
+## Setting up a new case in SARAS
+
+2D and 3D cases require separate executables of ``SARAS``.
+Once the dimensionality of the case is decided, ``SARAS`` has to compiled with appropriate compilation flags.
+The executable file may then be shifted to a work folder where it will be run.
+
+### Get the executable file
+
+Users on UNIX systems are encouraged to use the BASH script, ``compileSaras.sh``, in the ``./compile/`` folder to compile ``SARAS``.
+Immediately below the license header at the top of the script, there are 6 flags which the user can enable/disable.
+A flag can be enabled or disabled by uncommenting or commenting it respectively.
+
+It is best to disable the ``EXECUTE_AFTER_COMPILE`` flag so that script doesn't run the executable immediately after compilation.
+If the ``EXECUTE_AFTER_COMPILE`` flag is disabled, the ``PROC`` variable, which sets the number of processes to run the executable with, can be ignored.
+The user must enable or disable the ``PLANAR`` flag depending on whether the case is 2D or 3D respectively.
+
+The ``REAL_TYPE`` flag has the default value of ``DOUBLE``.
+This indicates that the executable will use double precision floating point numbers.
+To use single precision floats, set the flag to ``SINGLE``.
+The remaining two flags, ``TIME_RUN`` and ``TEST_RUN``, are used in special cases only and is best left to their default disabled states.
+
+Once the above shell script variables have been set, the script can be executed at the command line to compile ``SARAS``.
+If the compilation occurs without hiccups, an executable file named ``saras`` will appear in the root folder of the solver.
+
+### Set the parameters file
+
+When ``saras`` is executed, it will first read the case parameters from a YAML file named ``parameters.yaml``.
+The user must set these parameters appropriately before executing ``saras``.
+
+A sample ``parameters.yaml`` file is provided with the solver in the ``./input/`` folder.
+The parameters are grouped under 5 sections, viz., ``Program``, ``Mesh``, ``Parallel``, ``Solver`` and ``Multigrid``.
+
+* The main parameters that set the boundary conditions, inital conditions, forcing/source terms, etc. are found under the ``Program`` section.
+* Grid parameters like number of points, stretching parameter for non-uniform grids, etc. are found under the ``Mesh`` section.
+* ``Parallel`` section lets the user define how many MPI sub-domains to decompose the computational domain into, and the number of OpenMP threads to use.
+* Non-dimensional time-step, file write intervals, final non-dimensional time and so on are set in the ``Solver`` section.
+* Finally, ``Multigrid`` section lets the user tweak the parameters of the Geometric Multigrid solver used to solve the pressure Poisson equation.
+
+Each parameter has documentation written into the ``parameters.yaml`` file itself.
+
+> Additionally, the user can add custom boundary conditions to the ``boundary`` library.
+> The source files of the ``boundary`` library can be found in ``./lib/boundary`` folder of the solver.
+> Similarly, the user can add custom initial conditions to the ``initial`` library in the ``./lib/initial/`` folder,
+> and custom forcing/source terms to the ``force`` library in ``./lib/force/``.
+> All the source files in these libraries have extensive Doxygen documentation, and are written to be as self-explanatory as possible.
+
+### Running and processing data
+
+The folder where the executable ``saras`` will be run must contain two subfolders - ``./input/`` and ``./output/``.
+The ``parameters.yaml`` must be saved in ``./input/``, and the solver will write data into ``./output/``.
+
[…]
