### 2020-06-26 roshansamuel/saras#2 @roshansamuel
Hi,

Thank you for helping us improve the documentation of the solver. In the commit 147c174 to master, we have added a new section to the README.md file, titled [Setting up a new case in SARAS](https://github.com/roshansamuel/saras#setting-up-a-new-case-in-saras).

We hope that the instructions in this section will be sufficient for the user to write their own examples with SARAS. If you feel that this addresses the issue adequately, please mark this issue as closed.

Thanks!

Roshan
