### 2020-07-07 ad5735b
Expanded the summary section of paper.md with a statement of intent
