+++ resources/paper.bib
@@ -11,7 +11,7 @@
 }
 
 @article{Harlow:PF1965,
-    author = {Harlow,Francis H. and Welch,J. Eddie},
+    author = {Harlow, Francis H. and Welch, J. Eddie},
     title = {Numerical Calculation of Time-Dependent Viscous Incompressible Flow of Fluid with Free Surface},
     journal = {Phys. Fluids},
     volume = {8},
@@ -181,4 +181,68 @@
     doi={10.1017/9781316810019},
 	Title = {Energy transfers in Fluid Flows: Multiscale and Spectral Perspectives},
 	Year = {2019}
+}
+
+@article{Verma:SNCS2020,
+    author={Verma, Mahendra K. and Samuel, Roshan and Chatterjee, Soumyadeep and Bhattacharya, Shashwat and Asad, Ali},
+    title={Challenges in Fluid Flow Simulations Using Exascale Computing},
+    journal={SN Computer Science},
+    year={2020},
+    month={May},
+    volume={1},
+    number={3},
+    pages={178},
+    issn={2661-8907},
+    doi={10.1007/s42979-020-00184-1}
+}
+
+@article{Weller:1998FOAM,
+    title={A tensorial approach to computational continuum mechanics using object-oriented techniques},
+    author={Weller, Henry G. and Tabor, Gavin and Jasak, Hrvoje and Fureby, Christer},
+    journal={Computers in physics},
+    volume={12},
+    number={6},
+    pages={620--631},
+    year={1998},
+    doi = {10.1063/1.168744},
+    publisher={American Institute of Physics}
+}
+
+@article{Popinet:2003JCP,
+    title = "Gerris: a tree-based adaptive solver for the incompressible {E}uler equations in complex geometries",
+    journal = "Journal of Computational Physics",
+    volume = "190",
+    number = "2",
+    pages = "572 - 600",
+    year = "2003",
+    issn = "0021-9991",
+    doi = "10.1016/S0021-9991(03)00298-5",
+    url = "http://www.sciencedirect.com/science/article/pii/S0021999103002985",
+    author = "Stéphane Popinet"
+}
+
+@inbook{Palacios:2013AIAA,
+    author = {Palacios, Francisco and Alonso, Juan J. and Colonno, Michael R. and Aranake, Aniket C. and Campos, Alejandro and Copeland, Sean R. and Economon, Thomas D. and Lonkar, Amrita K. and Lukaczyk, Trent W. and Taylor, Thomas W. R.},
+    title = {Stanford {U}niversity {U}nstructured ({SU2}): {A}n open-source integrated computational environment for multi-physics simulation and design},
+    booktitle = {51st {AIAA} {A}erospace {S}ciences {M}eeting including the {N}ew {H}orizons {F}orum and {A}erospace {E}xposition},
+    chapter = {},
+    pages = {},
+    year = "2013",
+    doi = {10.2514/6.2013-287},
+    URL = {https://arc.aiaa.org/doi/abs/10.2514/6.2013-287},
+    eprint = {https://arc.aiaa.org/doi/pdf/10.2514/6.2013-287}
+}
+
+@article{Brandenburg:2002CPC,
+    title = "Hydromagnetic turbulence in computer simulations",
+    journal = "Computer Physics Communications",
+    volume = "147",
+    number = "1",
+    pages = "471 - 475",
+    year = "2002",
+    note = "Proceedings of the Europhysics Conference on Computational Physics Computational Modeling and Simulation of Complex Systems",
+    issn = "0010-4655",
+    doi = "10.1016/S0010-4655(02)00334-X",
+    url = "http://www.sciencedirect.com/science/article/pii/S001046550200334X",
+    author = "A. Brandenburg and W. Dobler"
 }
