### 2020-07-07 roshansamuel/saras#3 @roshansamuel
Hi,

Thank you very much for the helpful review comments. In the commit ad5735b, the [Summary](https://github.com/roshansamuel/saras/blob/master/paper.md#summary) section of the paper.md file has been expanded. We have added a brief explanation on intent of software. We also cite a recent paper which uses results from the SARAS code.

Although many popular CFD solvers now use finite-volume method or newer methods like LBM, finite-difference methods still offer comparatively higher accuracy on regular Cartesian grids. We wrote SARAS with this emphasis on accuracy and speed for academic problems.

If you feel that the expanded summary of the paper sufficiently explains the intent of the software, kindly mark this issue as closed.

Thanks!

Roshan
