### 2020-07-07 #2095 @roshansamuel
Hi @kyleniemeyer and @dlagrava, out of the three issues opened on the repository, [one](https://github.com/roshansamuel/saras/issues/1) was closed on Jun 4 following an update to the [README](https://github.com/roshansamuel/saras/blob/master/README.md). We have responded to [Issue 2](https://github.com/roshansamuel/saras/issues/2) by adding instructions on [how to set up a new case](https://github.com/roshansamuel/saras#setting-up-a-new-case-in-saras) in the README. To address [Issue 3](https://github.com/roshansamuel/saras/issues/3), we expanded the [Summary](https://github.com/roshansamuel/saras/blob/master/paper.md#summary) section of the paper with a statement of intent, and a brief comparison with other similar software.

Please let us know if all the open issues on the repository can now be closed.

Thanks!
