### 2020-07-08 #2095 @kyleniemeyer
Hi @roshansamuel, I think your changes have addressed @dlagrava's concerns, thanks!
