### 2020-08-05 roshansamuel/saras#3 @dlagrava
Hi Roshan,

I completely forgot about closing this one. I think the solution is good enough.

Thanks,
Daniel
