### 2020-10-20 #2095 @roshansamuel
Dear @kyleniemeyer ,

We are using data from simulations with SARAS in upcoming papers from our lab. It would be greatly helpful if the review of this software can be expedited. Once published, it would be easier for us to cite the solver in our papers.

Can something be done to bring the review back on track?

Thanks!
