### 2020-10-26 #2095 @kyleniemeyer
Hi @adam-m-jcbs, are you available to review this submission for JOSS?
