### 2021-01-16 #2095 @kyleniemeyer
@olgadoronina @nickwimer I have added you as reviewers—thank you for agreeing! I hope you can submit your reviews within the next two weeks. You should be able to edit the checklists above, and our review criteria are described in more detail at https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @roshansamuel, just to update you: I have found two new reviewers, and I hope we can have this wrapped up in the next few weeks—thanks for your patience.
