### 2021-02-05 #2095 @nickwimer
@roshansamuel Twice in the paper, SARAS has been compared to TARANG with statements:

1) > "TARANG has been shown to scale up to 196608 cores (Chatterjee et al., 2018), and SARAS has been designed with the goal of achieving similar scaling performance."

 and

2) > "To validate the accuracy of the finite difference scheme of SARAS, we compare the results of SARAS with those of a pseudo-spectral code TARANG, which has been benchmarked and scaled up to 196608 cores of Cray XC40, Shaheen II of KAUST (Chatterjee et al., 2018; Verma et al., 2013)."

Yet, I do not see any metrics of performance for SARAS presented in the paper. These statements (in particular, statement #2) imply that SARAS is also capable of such performance. A statement should be made in the paper making it clear that these goals of performance either have not been met, or remain untested.
