### 2021-02-05 #2095 @nickwimer
@roshansamuel, I am having several compilation errors on MacOS Mojave, Version 10.14.6. @olgadoronina has helped provide some useful compile flags to assist, but there are several outstanding issues.

Is it your intention to support the use of SARAS on MacOS, or is this strictly for LINUX? If you intend to support MacOS, I think more documentation would be helpful to assist in the widespread use of SARAS. If you are targeting only LINUX systems, I can work on installing this on a different system.
