### 2021-02-06 #2095 @roshansamuel
>
> Yet, I do not see any metrics of performance for SARAS presented in the paper. These statements (in particular, statement #2) imply that SARAS is also capable of such performance. A statement should be made in the paper making it clear that these goals of performance either have not been met, or remain untested.

Hi, @nickwimer, thank you for pointing this out. We haven't yet performed a scaling study of SARAS yet. We will add a line in the paper to clearly state this.

Since SARAS has a design similar to TARANG, and it requires lesser MPI communication than TARANG, we expect SARAS to scale as well as, if not better than TARANG. However, this is yet to be confirmed, and we will update the paper to clarify that.
