### 2021-02-09 #2095 @roshansamuel
>
>
> @roshansamuel, I am having several compilation errors on MacOS Mojave, Version 10.14.6. @olgadoronina has helped provide some useful compile flags to assist, but there are several outstanding issues.
>
> Is it your intention to support the use of SARAS on MacOS, or is this strictly for LINUX? If you intend to support MacOS, I think more documentation would be helpful to assist in the widespread use of SARAS. If you are targeting only LINUX systems, I can work on installing this on a different system.

Hi @nickwimer, we developed and tested SARAS on Linux with gcc. However, we have installed SARAS on a Mac system (Mojave) as well at some point. We had used gcc and not clang for this. Can you kindly give more details on the compilation errors you faced? We are working on a fresh install of SARAS on a Mac machine so that we can update the installation instructions accordingly.
