### 2021-02-09 8d87eb4
Updated multi-grid solver for faster convergence. Added test for MG solver in tests. Minor updates in comments
