+++ lib/poisson/poisson.cc
@@ -62,6 +62,8 @@
 poisson::poisson(const grid &mesh, const parser &solParam): mesh(mesh), inputParams(solParam) {
     int maxIndex = 15;
 
+    all = blitz::Range::all();
+
     mgSizeArray.resize(maxIndex);
     for (int i=0; i < maxIndex; i++) {
         mgSizeArray(i) = int(pow(2, i)) + 1;
@@ -74,189 +76,12 @@ poisson::poisson(const grid &mesh, const parser &solParam): mesh(mesh), inputPar
         strideValues(i) = int(pow(2, i));
     }
 
-    vLevel = 0;
-    maxCount = mesh.collocCoreSize(0)*mesh.collocCoreSize(1)*mesh.collocCoreSize(2);
-
 #ifdef TIME_RUN
-    solveTimeComp = 0.0;
-    solveTimeTran = 0.0;
     smothTimeComp = 0.0;
     smothTimeTran = 0.0;
 #endif
 }
 
-/**
- ********************************************************************************************************************************************
- * \brief   Function to initialize the arrays used in multi-grid
- *
- *          The memory required for various arrays in multi-grid solver are pre-allocated through this function.
- *          The function is called from within the constructor to perform this allocation once and for all.
- *          The arrays are initialized to 0.
- ********************************************************************************************************************************************
- */
-void poisson::initializeArrays() {
-    pressureData.resize(blitz::TinyVector<int, 3>(stagFull.ubound() - stagFull.lbound() + 1));
-    pressureData.reindexSelf(stagFull.lbound());
-    pressureData = 0.0;
-
-    smoothedPres.resize(blitz::TinyVector<int, 3>(stagFull.ubound() - stagFull.lbound() + 1));
-    smoothedPres.reindexSelf(stagFull.lbound());
-    smoothedPres = 0.0;
-
-    iteratorTemp.resize(blitz::TinyVector<int, 3>(stagFull.ubound() - stagFull.lbound() + 1));
-    iteratorTemp.reindexSelf(stagFull.lbound());
-    iteratorTemp = 0.0;
-
-    inputRHSData.resize(blitz::TinyVector<int, 3>(stagFull.ubound() - stagFull.lbound() + 1));
-    inputRHSData.reindexSelf(stagFull.lbound());
-    inputRHSData = 0.0;
-
-    residualData.resize(blitz::TinyVector<int, 3>(stagFull.ubound() - stagFull.lbound() + 1));
-    residualData.reindexSelf(stagFull.lbound());
-    residualData = 0.0;
-}
-
-/**
- ********************************************************************************************************************************************
- * \brief   Function to solve the poisson equation at the coarsest multi-grid level
- *
- *          This function operates exclusively at the lowest level of the multi-grid V-cycle.
- *          It uses the Jacobi iterative solver to solve the residual of the Poisson equation on the coarsest mesh.
- *          Note that the all calculations are performed assuming that the \ref vLevel variable is maximal when the function
- *          is being called.
- ********************************************************************************************************************************************
- */
-void poisson::solve() { };
-
-/**
- ********************************************************************************************************************************************
- * \brief   Function to perform prolongation on the array being solved
- *
- *          Prolongation makes the grid finer by averaging values at two adjacent nodes onto an intermediate point between them
- *          As a result, the number of points in the domain increases from \f$ 2^N + 1 \f$ at the input level to \f$ 2^{N+1} + 1 \f$.
- *          The vLevel variable is accordingly reduced by 1 to reflect this ascent by one step up the V-Cycle.
- ********************************************************************************************************************************************
- */
-void poisson::prolong() { };
-
-/**
- ********************************************************************************************************************************************
- * \brief   Function to perform smoothing operation on the input array
- *
[…]
