### 2021-02-09 c40d122
Updated README with detailed steps for installing SARAS on Linux and MacOS
