+++ README.md
@@ -1,7 +1,7 @@
 # Saras - Finite difference solver
 
 Saras is an OpenMP-MPI hybrid parallelized Navier-Stokes equation solver written in C++.
-It uses the finite-difference method for calculating spatial derivatives and parallelized geometric multigrid method for solving
+It uses the finite-difference method for calculating spatial derivatives and parallelized geometric multi-grid method for solving
 the pressure Poisson equation.
 
 All the source and library files for the Saras solver are contained in the following directories:
@@ -13,73 +13,154 @@ All the source and library files for the Saras solver are contained in the follo
 
 ## Installing SARAS
 
-To install ``SARAS``, you need to first clone the git repository into your local machine
+``SARAS`` relies on a few libraries for its calculations.
+Therefore the first step towards building ``SARAS`` is to install the following dependencies:
+
+* ``blitz`` - All array manipulations are performed using the Blitz++ library
+* ``cmake`` - Necessary to build the ``saras`` executable
+* ``mpich`` - For parallel computation using MPI
+* ``yaml`` - The input parameters are stored in a YAML file which needs to be parsed using yaml-cpp library.
+* ``hdf5`` - The output files are written in HDF5 format
+
+Packages like ``cmake``, ``mpich`` and ``hdf5`` can be installed from the OS package manager,
+while ``yaml`` and ``blitz`` can be downloaded and installed manually.
+
+However, we provide below an alternative series of steps that will install all the libraries in the user's home directory.
+This method has three advantages:
+
+* It does not require administrator (``sudo``) privileges,
+* It will not disturb pre-existing packages already installed on the system
+* The method has been tested on both Linux (Ubuntu 14.04 and above), and MacOS (Mojave)
+
+### Download all the dependencies
+
+All the required packages can be downloaded from our [lab website](http://turbulencehub.org), and installed over a terminal.
+It is advisable to create a temporary directory where the packages can be downloaded and extracted.
+After navigating to the temporary directory, download the packages using ``wget`` on Linux, or ``curl`` on MacOS.
+For instance, on MacOS, the following lines will download ``CMake``, ``Blitz++``, ``yaml-cpp``, ``MPICH`` and ``HDF5`` packages respectively:
+
+`curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/cmake-2.8.12.tar.gz`
+`curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/blitz-1.0.1.tar.gz`
+`curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/yaml-cpp-release-0.3.0.tar.gz`
+`curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/mpich-3.1.3.tar.gz`
+`curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/hdf5-1.8.20.tar.bz2`
+
+On Linux, please replace ``curl -O`` with ``wget``.
+If you are installing on a remote machine which doesn't have direct internet access,
+the packages can be downloaded using the links listed above, and transferred over ssh or ftp.
+
+### Extract all the packages
+
+The ``tar`` command (available by default on both Linux and MacOS) can be used to extract the above archives.
+For instance, ``CMake`` can be extracted by the command:
+
+`tar -xf cmake-2.8.12.tar.gz`
+
+On MacOS, if the above command fails with the error ``Failed to set default locale``, the $LANG environment variable has to be set as shown below:
+
+`export LANG=en_US.UTF-8`
+
+Once all the downloaded packages have been extracted, create an install location in the user's home folder if it doesn't exist already.
+Usually, packages are installed in ``$HOME/local/`` directory.
[…]
