### 2021-02-09 83d19b9
Updated paper and eliminated a compile-time warning in the io library
