+++ README.md
@@ -39,15 +39,13 @@ It is advisable to create a temporary directory where the packages can be downlo
 After navigating to the temporary directory, download the packages using ``wget`` on Linux, or ``curl`` on MacOS.
 For instance, on MacOS, the following lines will download ``CMake``, ``Blitz++``, ``yaml-cpp``, ``MPICH`` and ``HDF5`` packages respectively:
 
-`curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/cmake-2.8.12.tar.gz`
-
-`curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/blitz-1.0.1.tar.gz`
-
-`curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/yaml-cpp-release-0.3.0.tar.gz`
-
-`curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/mpich-3.1.3.tar.gz`
-
-`curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/hdf5-1.8.20.tar.bz2`
+```
+curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/cmake-2.8.12.tar.gz
+curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/blitz-1.0.1.tar.gz
+curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/yaml-cpp-release-0.3.0.tar.gz
+curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/mpich-3.1.3.tar.gz
+curl -O https://turbulencehub.org/wp-content/uploads/Download_Files/hdf5-1.8.20.tar.bz2
+```
 
 On Linux, please replace ``curl -O`` with ``wget``.
 If you are installing on a remote machine which doesn't have direct internet access,
@@ -72,11 +70,11 @@ The next steps will assume that the folder ``local/`` exists in the user's home 
 
 Navigate to the folder created by extracting the ``CMake`` package, configure the installation script, and install the package:
 
-`cd cmake-2.8.12/`
-
-`./configure --prefix=$HOME/local`
-
-`make -j4 install`
+```
+cd cmake-2.8.12/
+./configure --prefix=$HOME/local
+make -j4 install
+```
 
 Note that the option ``-j4`` given to ``make`` will use 4 cores of your system.
 If more cores are available, the process can be speeded-up by specifying a higher number.
@@ -87,19 +85,15 @@ Once ``CMake`` has been installed in ``$HOME/local`` it should be made available
 For this, the path variables have to be updated to let the system know that ``cmake`` is installed in ``$HOME/local``.
 Accordingly, set the following environment variables:
 
-`export PATH=$HOME/local/bin:$PATH`
-
-`export PKG_CONFIG_PATH=$HOME/local/lib/pkgconfig:$PKG_CONFIG_PATH`
-
-`export HDF5_ROOT=$HOME/local`
-
-`export CPATH=$HOME/local/include/:$CPATH`
-
-`export LD_LIBRARY_PATH=$HOME/local/lib:$LD_LIBRARY_PATH`
-
-`export LIBRARY_PATH=$HOME/local/lib:$LIBRARY_PATH`
-
-`export MANPATH=$HOME/local/share/man/:$MANPATH`
+```
+export PATH=$HOME/local/bin:$PATH
+export PKG_CONFIG_PATH=$HOME/local/lib/pkgconfig:$PKG_CONFIG_PATH
+export HDF5_ROOT=$HOME/local
+export CPATH=$HOME/local/include/:$CPATH
[…]
