### 2021-02-09 #2095 @roshansamuel
Hi @nickwimer, we have updated the README with clearer instructions to install SARAS on both Linux and MacOS. The steps worked with both homebrew-gcc as well as clang on MacOS. We have also added the line "_We are planning to conduct a scaling analysis of SARAS and compare its performance with TARANG in the near future._" in the paper (Summary, Para 3, Line 3) to clarify that the scaling is yet to be studied.

Kindly let me know if the new installation instructions help to avoid the compilation errors you had faced. Thanks!
