### 2021-02-10 #2095 @nickwimer
@roshansamuel Thank you for updating the README with instructions for MacOS. I am will work on getting SARAS installed today/tomorrow and I will let you know if I run into any more issues.
