### 2021-02-12 #2095 @nickwimer
@roshansamuel Still running into compile issues. Below is the current error message:

> -- The parent directory is /Users/nwimer/Documents/Reviews/saras
-- Install folder is /Users/nwimer/Documents/Reviews/saras/compile
-- All headers are in /Users/nwimer/Documents/Reviews/saras/lib
-- Compiler flag for OpenMP is -Xclang -fopenmp -Wno-unused-command-line-argument
-- PLANAR flag not set. Compiling for default 3D run
-- Compiling Saras for 3D simulations
-- Compiling Saras to solve with double precision calculations
-- Configuring done
-- Generating done
-- Build files have been written to: /Users/nwimer/Documents/Reviews/saras/compile/build
[  2%] [  9%] [ 12%] [ 14%] [ 31%] [ 39%] Built target initial
Built target field
[ 41%] Built target reader
Built target tseries
[ 43%] Built target poisson
[ 48%] Built target parser
Built target writer
[ 51%] Built target grid
[ 73%] [ 82%] Built target parallel
Built target probes
Built target boundary
Built target force
Linking CXX executable ../../../../saras
Undefined symbols for architecture x86_64:
  "___kmpc_for_static_fini", referenced from:
      _.omp_outlined. in hydro_d2.cc.o
      _.omp_outlined..1 in hydro_d2.cc.o
      _.omp_outlined..3 in hydro_d2.cc.o
      _.omp_outlined..4 in hydro_d2.cc.o
      _.omp_outlined. in hydro_d3.cc.o
      _.omp_outlined..1 in hydro_d3.cc.o
      _.omp_outlined..3 in hydro_d3.cc.o
      ...
  "___kmpc_for_static_init_4", referenced from:
      _.omp_outlined. in hydro_d2.cc.o
      _.omp_outlined..1 in hydro_d2.cc.o
      _.omp_outlined..3 in hydro_d2.cc.o
      _.omp_outlined..4 in hydro_d2.cc.o
      _.omp_outlined. in hydro_d3.cc.o
      _.omp_outlined..1 in hydro_d3.cc.o
      _.omp_outlined..3 in hydro_d3.cc.o
      ...
  "___kmpc_fork_call", referenced from:
      hydro_d2::solveVx() in hydro_d2.cc.o
      hydro_d2::solveVz() in hydro_d2.cc.o
      hydro_d3::solveVx() in hydro_d3.cc.o
      hydro_d3::solveVy() in hydro_d3.cc.o
      hydro_d3::solveVz() in hydro_d3.cc.o
      scalar_d2::solveVx() in scalar_d2.cc.o
      scalar_d2::solveVz() in scalar_d2.cc.o
      ...
  "___kmpc_global_thread_num", referenced from:
      hydro_d2::solveVx() in hydro_d2.cc.o
      _.omp_outlined. in hydro_d2.cc.o
      _.omp_outlined..1 in hydro_d2.cc.o
      hydro_d2::solveVz() in hydro_d2.cc.o
      _.omp_outlined..3 in hydro_d2.cc.o
      _.omp_outlined..4 in hydro_d2.cc.o
      hydro_d3::solveVx() in hydro_d3.cc.o
      ...
  "___kmpc_push_num_threads", referenced from:
      hydro_d2::solveVx() in hydro_d2.cc.o
      hydro_d2::solveVz() in hydro_d2.cc.o
      hydro_d3::solveVx() in hydro_d3.cc.o
      hydro_d3::solveVy() in hydro_d3.cc.o
      hydro_d3::solveVz() in hydro_d3.cc.o
      scalar_d2::solveVx() in scalar_d2.cc.o
      scalar_d2::solveVz() in scalar_d2.cc.o
      ...
ld: symbol(s) not found for architecture x86_64
clang: error: linker command failed with exit code 1 (use -v to see invocation)
make[2]: *** [../../saras] Error 1
make[1]: *** [src/solvers/CMakeFiles/saras.dir/all] Error 2
make: *** [all] Error 2
