### 2021-02-12 #2095 @olgadoronina
Hi @roshansamuel, thank you for updating the installation instructions. I was able to install all the packages with just a few changes:
- blitz++
	```
 	% make -j4 install
	Making install in bin
	make[2]: Nothing to be done for `install-exec-am'.
	make[2]: Nothing to be done for `install-data-am'.
	Making install in blitz
	Making install in generate
	/Library/Developer/CommandLineTools/usr/bin/make  generate-headers
	python genstencils.py ../../blitz/array/stencil-classes.cc
	  File "genstencils.py", line 9
		print "Generating file %s"%sys.argv[1]
			  ^
	SyntaxError: Missing parentheses in call to 'print'. Did you mean print("Generating file %s"%sys.argv[1])?
	make[3]: *** [../../blitz/array/stencil-classes.cc] Error 1
	make[2]: *** [all-local] Error 2
	make[1]: *** [install-recursive] Error 1
	make: *** [install-recursive] Error 1
	```
	Fix: activate conda environment with python2
- MPICH (`./configure --prefix=$HOME/local`)
	```
	configure: error: The Fortran compiler gfortran will not compile files that call the same routine with arguments of different types.
	```
	Fix: `export FFLAGS="-w -fallow-argument-mismatch -O2"`

- hdf5 installation
	```
	cache_api.c:261:16: error: implicit declaration of function ‘resize_configs_are_equal’ is invalid in C99  [-Werror,-Wimplicit-function-declaration]  if ( ! resize_configs_are_equal(&default_auto_size_ctl
	```
         Fix: `export CFLAGS=-Wno-error=implicit-function-declaration`

I was able to compile SARAS without errors using compileSaras.sh script, but it seems that compiler flag for OpenMP wasn't set up and test run fails. See output for compileSaras.sh below. Could you help me fix it?
>% bash compileSaras.sh
-- The parent directory is /Users/odoronin/saras-master
-- Install folder is /Users/odoronin/saras-master/compile
-- All headers are in /Users/odoronin/saras-master/lib
-- Try OpenMP C flag = [-fopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [/openmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [-Qopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [-openmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [ ]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [-xopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [+Oopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [-qsmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [-mp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [-fopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [/openmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [-Qopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [-openmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [ ]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [-xopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [+Oopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [-qsmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [-mp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Could NOT find OpenMP (missing:  OpenMP_C_FLAGS OpenMP_CXX_FLAGS)
-- Compiler flag for OpenMP is
-- PLANAR flag not set. Compiling for default 3D run
-- Compiling Saras for 3D simulations
-- Compiling Saras to solve with double precision calculations
-- Configuring done
-- Generating done
-- Build files have been written to: /Users/odoronin/saras-master/compile/build
[  4%] [  7%] [  9%] [ 12%] [ 14%] [ 19%] Built target grid
[ 31%] Built target parallel
Built target tseries
Built target reader
[ 58%] [ 60%] Built target parser
Built target writer
Built target field
Built target initial
[ 63%] Built target boundary
[ 75%] [ 82%] Built target probes
Built target force
Built target poisson
Linking CXX executable ../../../../saras
[100%] Built target saras
