### 2021-02-13 #2095 @roshansamuel
Hi @olgadoronina, thank you very much for suggesting the fixes for installing different libraries. I will add them to the installation instructions.

According to the output you have posted, the executable seems to have been built. You can execute it with mpirun and it will run, although it will not use OpenMP as it seems to not have detected it. Is `libomp-dev` installed on your system? I didn't include installation of OpenMP libraries in the instructions since the solver still ran on different machines without them.

In order to check if the solver is really compiling right, can you please run `testSaras.sh` script inside the `tests/` folder? Please let me know if that test fails.
