### 2021-02-13 #2095 @roshansamuel
Hi @nickwimer. From the error message, it looks like the issue is with OpenMP. When I tested the installation on 2 machines, (Ubuntu and Mac), I had skipped installation of OpenMP since the solver still ran on those systems without it. Can you please let me know if OpenMP is installed on your system? I will update the installation instructions to include OpenMP libraries in the meantime.
