### 2021-02-15 #2095 @olgadoronina
Hi @roshansamuel, yes, the executable was built, but the test failed. See the output below. It seems like it can't run without OpenMP. I have libomp installed through homebrew, I'm not sure why `find_package (OpenMP)` doesn't return anything. Do I need to install something else?
>% bash testSaras.sh
-- The C compiler identification is Clang 12.0.0
-- The CXX compiler identification is Clang 12.0.0
-- Check for working C compiler: /Users/odoronin/local/bin/mpicc
-- Check for working C compiler: /Users/odoronin/local/bin/mpicc -- works
-- Detecting C compiler ABI info
-- Detecting C compiler ABI info - done
-- Check for working CXX compiler: /Users/odoronin/local/bin/mpicxx
-- Check for working CXX compiler: /Users/odoronin/local/bin/mpicxx -- works
-- Detecting CXX compiler ABI info
-- The parent directory is /Users/odoronin/saras-master
-- Install folder is /Users/odoronin/saras-master/tests
-- All headers are in /Users/odoronin/saras-master/lib
-- Try OpenMP C flag = [-fopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [/openmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [-Qopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [-openmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [ ]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [-xopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [+Oopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [-qsmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP C flag = [-mp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [-fopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [/openmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [-Qopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [-openmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [ ]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [-xopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [+Oopenmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [-qsmp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Try OpenMP CXX flag = [-mp]
-- Performing Test OpenMP_FLAG_DETECTED
-- Performing Test OpenMP_FLAG_DETECTED - Failed
-- Could NOT find OpenMP (missing:  OpenMP_C_FLAGS OpenMP_CXX_FLAGS)
-- Compiler flag for OpenMP is
-- Compiling Saras for 2D simulations
-- Compiling Saras to solve with double precision calculations
-- Configuring done
-- Generating done
-- Build files have been written to: /Users/odoronin/saras-master/tests/build
Scanning dependencies of target field
Scanning dependencies of target grid
Scanning dependencies of target poisson
Scanning dependencies of target parallel
Scanning dependencies of target parser
Scanning dependencies of target probes
Scanning dependencies of target reader
Scanning dependencies of target tseries
[  7%] [  7%] [  7%] [ 12%] [ 12%] [ 14%] [ 17%] [ 19%] Building CXX object lib/io/CMakeFiles/parser.dir/parser.cc.o
Building CXX object lib/io/CMakeFiles/probes.dir/probes.cc.o
Building CXX object lib/grid/CMakeFiles/grid.dir/grid.cc.o
Building CXX object lib/parallel/CMakeFiles/parallel.dir/parallel.cc.o
Building CXX object lib/io/CMakeFiles/reader.dir/reader.cc.o
Building CXX object lib/io/CMakeFiles/tseries.dir/tseries.cc.o
Building CXX object lib/poisson/CMakeFiles/poisson.dir/poisson.cc.o
Building CXX object lib/field/CMakeFiles/field.dir/field.cc.o
In file included from /Users/odoronin/saras-master/lib/io/tseries.cc:44:
/Users/odoronin/saras-master/lib/io/tseries.h:64:13: warning: private field 'yLow' is not used [-Wunused-private-field]
        int yLow, yTop;
            ^
/Users/odoronin/saras-master/lib/io/tseries.h:64:19: warning: private field 'yTop' is not used [-Wunused-private-field]
        int yLow, yTop;
                  ^
[ 21%] Building CXX object lib/parallel/CMakeFiles/parallel.dir/mpidata.cc.o
Linking CXX static library libreader.a
[ 21%] Built target reader
[ 24%] Building CXX object lib/field/CMakeFiles/field.dir/sfield.cc.o
Linking CXX static library libprobes.a
[ 24%] Built target probes
[ 26%] Building CXX object lib/field/CMakeFiles/field.dir/vfield.cc.o
2 warnings generated.
Linking CXX static library libtseries.a
[ 26%] Built target tseries
[ 29%] Building CXX object lib/poisson/CMakeFiles/poisson.dir/poisson_d2.cc.o
Linking CXX static library libparser.a
[ 29%] Built target parser
[ 31%] Building CXX object lib/field/CMakeFiles/field.dir/plainsf.cc.o
Scanning dependencies of target writer
[ 34%] Building CXX object lib/io/CMakeFiles/writer.dir/writer.cc.o
Linking CXX static library libparallel.a
Linking CXX static library libgrid.a
[ 34%] Built target parallel
[ 36%] Building CXX object lib/field/CMakeFiles/field.dir/plainvf.cc.o
[ 36%] Built target grid
[ 39%] Building CXX object lib/poisson/CMakeFiles/poisson.dir/poisson_d3.cc.o
[ 41%] Building CXX object lib/field/CMakeFiles/field.dir/derivative.cc.o
In file included from /Users/odoronin/saras-master/lib/field/plainvf.cc:43:
/Users/odoronin/saras-master/lib/field/plainvf.h:51:21: warning: private field 'gridData' is not used [-Wunused-private-field]
        const grid &gridData;
                    ^
Linking CXX static library libwriter.a
[ 41%] Built target writer
Scanning dependencies of target boundary
[ 43%] Scanning dependencies of target initial
Building CXX object lib/boundary/CMakeFiles/boundary.dir/boundary.cc.o
[ 46%] Building CXX object lib/initial/CMakeFiles/initial.dir/initial.cc.o
Scanning dependencies of target force
[ 48%] Building CXX object lib/force/CMakeFiles/force.dir/force.cc.o
[ 51%] [ 53%] Building CXX object lib/initial/CMakeFiles/initial.dir/taylorGreen.cc.o
Building CXX object lib/boundary/CMakeFiles/boundary.dir/dirichletCC.cc.o
1 warning generated.
[ 56%] Building CXX object lib/force/CMakeFiles/force.dir/coriolisForce.cc.o
[ 58%] Building CXX object lib/initial/CMakeFiles/initial.dir/channelSine.cc.o
[ 60%] Building CXX object lib/force/CMakeFiles/force.dir/buoyantForce.cc.o
[ 63%] Building CXX object lib/boundary/CMakeFiles/boundary.dir/dirichletFC.cc.o
[ 65%] Building CXX object lib/initial/CMakeFiles/initial.dir/channelRand.cc.o
[ 68%] Building CXX object lib/force/CMakeFiles/force.dir/rotatingConv.cc.o
Linking CXX static library libfield.a
[ 68%] Built target field
[ 70%] Building CXX object lib/boundary/CMakeFiles/boundary.dir/periodicCC.cc.o
Linking CXX static library libpoisson.a
[ 70%] Built target poisson
[ 73%] Building CXX object lib/force/CMakeFiles/force.dir/randomForcing.cc.o
[ 75%] Building CXX object lib/boundary/CMakeFiles/boundary.dir/periodicFC.cc.o
[ 78%] Building CXX object lib/boundary/CMakeFiles/boundary.dir/neumannCC.cc.o
[ 80%] Building CXX object lib/boundary/CMakeFiles/boundary.dir/neumannFC.cc.o
Linking CXX static library libinitial.a
[ 80%] Built target initial
[ 82%] Building CXX object lib/boundary/CMakeFiles/boundary.dir/hotPlateCC.cc.o
In file included from /Users/odoronin/saras-master/lib/boundary/neumannCC.cc:43:
/Users/odoronin/saras-master/lib/boundary/boundary.h:183:20: warning: private field 'fieldValue' is not used
      [-Wunused-private-field]
        const real fieldValue;
                   ^
In file included from /Users/odoronin/saras-master/lib/boundary/neumannFC.cc:43:
/Users/odoronin/saras-master/lib/boundary/boundary.h:200:20: warning: private field 'fieldValue' is not used
      [-Wunused-private-field]
        const real fieldValue;
                   ^
Linking CXX static library libforce.a
[ 82%] Built target force
1 warning generated.
1 warning generated.
Linking CXX static library libboundary.a
[ 82%] Built target boundary
Scanning dependencies of target saras
[ 85%] [ 87%] [ 90%] [ 92%] [ 95%] [ 97%] [100%] Building CXX object src/solvers/CMakeFiles/saras.dir/hydro/hydro.cc.o
Building CXX object src/solvers/CMakeFiles/saras.dir/main.cc.o
Building CXX object src/solvers/CMakeFiles/saras.dir/hydro/hydro_d2.cc.o
Building CXX object src/solvers/CMakeFiles/saras.dir/hydro/hydro_d3.cc.o
Building CXX object src/solvers/CMakeFiles/saras.dir/scalar/scalar.cc.o
Building CXX object src/solvers/CMakeFiles/saras.dir/scalar/scalar_d2.cc.o
Building CXX object src/solvers/CMakeFiles/saras.dir/scalar/scalar_d3.cc.o
Linking CXX executable ../../../../saras
[100%] Built target saras
Fatal error in MPI_Init: Other MPI error, error stack:
MPIR_Init_thread(498)..............:
MPID_Init(187).....................: channel initialization failed
MPIDI_CH3_Init(89).................:
MPID_nem_init(320).................:
MPID_nem_tcp_init(171).............:
MPID_nem_tcp_get_business_card(418):
MPID_nem_tcp_init(377).............: gethostbyname failed, odoronin-35746s (errno 1)
Fatal error in MPI_Init: Other MPI error, error stack:
MPIR_Init_thread(498)..............:
MPID_Init(187).....................: channel initialization failed
MPIDI_CH3_Init(89).................:
MPID_nem_init(320).................:
MPID_nem_tcp_init(171).............:
MPID_nem_tcp_get_business_card(418):
MPID_nem_tcp_init(377).............: gethostbyname failed, odoronin-35746s (errno 1)
Fatal error in MPI_Init: Other MPI error, error stack:
MPIR_Init_thread(498)..............:
MPID_Init(187).....................: channel initialization failed
MPIDI_CH3_Init(89).................:
MPID_nem_init(320).................:
MPID_nem_tcp_init(171).............:
MPID_nem_tcp_get_business_card(418):
MPID_nem_tcp_init(377).............: gethostbyname failed, odoronin-35746s (errno 1)
Fatal error in MPI_Init: Other MPI error, error stack:
MPIR_Init_thread(498)..............:
MPID_Init(187).....................: channel initialization failed
MPIDI_CH3_Init(89).................:
MPID_nem_init(320).................:
MPID_nem_tcp_init(171).............:
MPID_nem_tcp_get_business_card(418):
MPID_nem_tcp_init(377).............: gethostbyname failed, odoronin-35746s (errno 1)
===================================================================================
=   BAD TERMINATION OF ONE OF YOUR APPLICATION PROCESSES
=   PID 22525 RUNNING AT odoronin-35746s
=   EXIT CODE: 1
=   CLEANING UP REMAINING PROCESSES
=   YOU CAN IGNORE THE BELOW CLEANUP MESSAGES
===================================================================================
