### 2021-02-25 daf432b
Fix for OpenMP multi-threading issues on Mac systems
