+++ src/solvers/hydro/hydro_d3.cc
@@ -372,7 +372,6 @@ void hydro_d3::solveVx() {
     real maxError = 0.0;
 
     while (true) {
-#pragma omp parallel for num_threads(inputParams.nThreads) default(none)
         for (int iX = V.Vx.fBulk.lbound(0); iX <= V.Vx.fBulk.ubound(0); iX++) {
             for (int iY = V.Vx.fBulk.lbound(1); iY <= V.Vx.fBulk.ubound(1); iY++) {
                 for (int iZ = V.Vx.fBulk.lbound(2); iZ <= V.Vx.fBulk.ubound(2); iZ++) {
@@ -391,7 +390,6 @@ void hydro_d3::solveVx() {
 
         imposeUBCs();
 
-#pragma omp parallel for num_threads(inputParams.nThreads) default(none)
         for (int iX = V.Vx.fBulk.lbound(0); iX <= V.Vx.fBulk.ubound(0); iX++) {
             for (int iY = V.Vx.fBulk.lbound(1); iY <= V.Vx.fBulk.ubound(1); iY++) {
                 for (int iZ = V.Vx.fBulk.lbound(2); iZ <= V.Vx.fBulk.ubound(2); iZ++) {
@@ -429,7 +427,6 @@ void hydro_d3::solveVy() {
     real maxError = 0.0;
 
     while (true) {
-#pragma omp parallel for num_threads(inputParams.nThreads) default(none)
         for (int iX = V.Vy.fBulk.lbound(0); iX <= V.Vy.fBulk.ubound(0); iX++) {
             for (int iY = V.Vy.fBulk.lbound(1); iY <= V.Vy.fBulk.ubound(1); iY++) {
                 for (int iZ = V.Vy.fBulk.lbound(2); iZ <= V.Vy.fBulk.ubound(2); iZ++) {
@@ -448,7 +445,6 @@ void hydro_d3::solveVy() {
 
         imposeVBCs();
 
-#pragma omp parallel for num_threads(inputParams.nThreads) default(none)
         for (int iX = V.Vy.fBulk.lbound(0); iX <= V.Vy.fBulk.ubound(0); iX++) {
             for (int iY = V.Vy.fBulk.lbound(1); iY <= V.Vy.fBulk.ubound(1); iY++) {
                 for (int iZ = V.Vy.fBulk.lbound(2); iZ <= V.Vy.fBulk.ubound(2); iZ++) {
@@ -486,7 +482,6 @@ void hydro_d3::solveVz() {
     real maxError = 0.0;
 
     while (true) {
-#pragma omp parallel for num_threads(inputParams.nThreads) default(none)
         for (int iX = V.Vz.fBulk.lbound(0); iX <= V.Vz.fBulk.ubound(0); iX++) {
             for (int iY = V.Vz.fBulk.lbound(1); iY <= V.Vz.fBulk.ubound(1); iY++) {
                 for (int iZ = V.Vz.fBulk.lbound(2); iZ <= V.Vz.fBulk.ubound(2); iZ++) {
@@ -505,7 +500,6 @@ void hydro_d3::solveVz() {
 
         imposeWBCs();
 
-#pragma omp parallel for num_threads(inputParams.nThreads) default(none)
         for (int iX = V.Vz.fBulk.lbound(0); iX <= V.Vz.fBulk.ubound(0); iX++) {
             for (int iY = V.Vz.fBulk.lbound(1); iY <= V.Vz.fBulk.ubound(1); iY++) {
                 for (int iZ = V.Vz.fBulk.lbound(2); iZ <= V.Vz.fBulk.ubound(2); iZ++) {
