### 2021-02-25 51809d8
Updated README with fixes for some installation glitches
