### 2021-02-25 #2095 @roshansamuel
Hi @olgadoronina , @nickwimer, I have updated the solver and the README. I have added the fixes suggested by @olgadoronina for installation of some of the dependencies. Please let me know if the solver is now installing and running correctly.

Thanks!
