### 2021-03-07 #2095 @olgadoronina
Hi @roshansamuel, thank you for your updates, I was able to compile solver by following you installation instructions, but when I run the testSaras.sh, I get the following output:
``` tests % bash testSaras.sh
-- The parent directory is /Users/odoronin/saras-master
-- Install folder is /Users/odoronin/saras-master/tests
-- All headers are in /Users/odoronin/saras-master/lib
-- Compiling Saras for 2D simulations
-- Compiling Saras to solve with double precision calculations
-- Configuring done
-- Generating done
-- Build files have been written to: /Users/odoronin/saras-master/tests/build
[  4%] [  4%] [  7%] [ 21%] [ 24%] [ 26%] Built target tseries
[ 31%] Built target probes
Built target parser
[ 39%] Built target field
Built target grid
Built target reader
Built target parallel
Built target poisson
[ 41%] [ 60%] [ 73%] [ 82%] Built target writer
Built target boundary
Built target force
Built target initial
[100%] Built target saras
Fatal error in MPI_Init: Other MPI error, error stack:
MPIR_Init_thread(498)..............:
MPID_Init(187).....................: channel initialization failed
MPIDI_CH3_Init(89).................:
MPID_nem_init(320).................:
MPID_nem_tcp_init(171).............:
MPID_nem_tcp_get_business_card(418):
MPID_nem_tcp_init(377).............: gethostbyname failed, odoronin-35746s (errno 1)
Fatal error in MPI_Init: Other MPI error, error stack:
MPIR_Init_thread(498)..............:
MPID_Init(187).....................: channel initialization failed
MPIDI_CH3_Init(89).................:
MPID_nem_init(320).................:
MPID_nem_tcp_init(171).............:
MPID_nem_tcp_get_business_card(418):
MPID_nem_tcp_init(377).............: gethostbyname failed, odoronin-35746s (errno 1)
Fatal error in MPI_Init: Other MPI error, error stack:
MPIR_Init_thread(498)..............:
MPID_Init(187).....................: channel initialization failed
MPIDI_CH3_Init(89).................:
MPID_nem_init(320).................:
MPID_nem_tcp_init(171).............:
MPID_nem_tcp_get_business_card(418):
MPID_nem_tcp_init(377).............: gethostbyname failed, odoronin-35746s (errno 1)
Fatal error in MPI_Init: Other MPI error, error stack:
MPIR_Init_thread(498)..............:
MPID_Init(187).....................: channel initialization failed
MPIDI_CH3_Init(89).................:
MPID_nem_init(320).................:
MPID_nem_tcp_init(171).............:
MPID_nem_tcp_get_business_card(418):
MPID_nem_tcp_init(377).............: gethostbyname failed, odoronin-35746s (errno 1)

===================================================================================
=   BAD TERMINATION OF ONE OF YOUR APPLICATION PROCESSES
=   PID 75888 RUNNING AT odoronin-35746s
=   EXIT CODE: 1
=   CLEANING UP REMAINING PROCESSES
=   YOU CAN IGNORE THE BELOW CLEANUP MESSAGES
===================================================================================
validate_ldc.py:86: YAMLLoadWarning: calling yaml.load() without Loader=... is deprecated, as the default Loader is unsafe. Please read https://msg.pyyaml.org/load for full details.
  yamlData = yl.load(yamlFile)
Could not open file output/Soln_0030.0000.h5
```

If I run compileSaras.sh with `TEST_RUN="TEST_RUN"` I get the following errors:
```
% bash compileSaras.sh
-- The C compiler identification is Clang 12.0.0
-- The CXX compiler identification is Clang 12.0.0
-- Check for working C compiler: /Users/odoronin/local/bin/mpicc
-- Check for working C compiler: /Users/odoronin/local/bin/mpicc -- works
-- Detecting C compiler ABI info
-- Detecting C compiler ABI info - done
-- Check for working CXX compiler: /Users/odoronin/local/bin/mpicxx
-- Check for working CXX compiler: /Users/odoronin/local/bin/mpicxx -- works
-- Detecting CXX compiler ABI info
-- The parent directory is /Users/odoronin/saras-master
-- Install folder is /Users/odoronin/saras-master/compile
-- All headers are in /Users/odoronin/saras-master/lib
-- PLANAR flag not set. Compiling for default 3D run
-- Compiling Saras for 3D simulations
-- Compiling Saras for running unit tests
-- Configuring done
-- Generating done
-- Build files have been written to: /Users/odoronin/saras-master/compile/build
Scanning dependencies of target poisson
Scanning dependencies of target parser
Scanning dependencies of target field
Scanning dependencies of target grid
Scanning dependencies of target parallel
Scanning dependencies of target reader
Scanning dependencies of target probes
Scanning dependencies of target tseries
Scanning dependencies of target initial
Scanning dependencies of target writer
Scanning dependencies of target boundary
Scanning dependencies of target force
[  2%] [  4%] [  6%] [ 11%] [ 11%] [ 13%] [ 15%] [ 18%] [ 20%] [ 25%] [ 25%] [ 27%] [ 29%] [ 31%] Building CXX object lib/io/CMakeFiles/parser.dir/parser.cc.o
Building CXX object lib/grid/CMakeFiles/grid.dir/grid.cc.o
Building CXX object lib/parallel/CMakeFiles/parallel.dir/parallel.cc.o
[ 34%] [ 36%] Building CXX object lib/io/CMakeFiles/reader.dir/reader.cc.o
Building CXX object lib/parallel/CMakeFiles/parallel.dir/mpidata.cc.o
Building CXX object lib/io/CMakeFiles/probes.dir/probes.cc.o
Building CXX object lib/io/CMakeFiles/writer.dir/writer.cc.o
Building CXX object lib/io/CMakeFiles/tseries.dir/tseries.cc.o
Building CXX object lib/poisson/CMakeFiles/poisson.dir/poisson.cc.o
Building CXX object lib/poisson/CMakeFiles/poisson.dir/poisson_d2.cc.o
Building CXX object lib/initial/CMakeFiles/initial.dir/initial.cc.o
Building CXX object lib/field/CMakeFiles/field.dir/field.cc.o
Building CXX object lib/poisson/CMakeFiles/poisson.dir/poisson_d3.cc.o
Building CXX object lib/initial/CMakeFiles/initial.dir/taylorGreen.cc.o
Building CXX object lib/boundary/CMakeFiles/boundary.dir/boundary.cc.o
Building CXX object lib/force/CMakeFiles/force.dir/force.cc.o
[ 38%] Building CXX object lib/boundary/CMakeFiles/boundary.dir/dirichletCC.cc.o
[ 40%] [ 43%] Building CXX object lib/force/CMakeFiles/force.dir/coriolisForce.cc.o
Building CXX object lib/field/CMakeFiles/field.dir/sfield.cc.o
[ 45%] Building CXX object lib/boundary/CMakeFiles/boundary.dir/dirichletFC.cc.o
Linking CXX static library libparallel.a
[ 45%] Built target parallel
Linking CXX static library libreader.a
[ 47%] Building CXX object lib/initial/CMakeFiles/initial.dir/channelSine.cc.o
[ 50%] [ 50%] Building CXX object lib/force/CMakeFiles/force.dir/buoyantForce.cc.o
Built target reader
[ 52%] Building CXX object lib/field/CMakeFiles/field.dir/vfield.cc.o
Linking CXX static library libprobes.a
[ 52%] Built target probes
[ 54%] Building CXX object lib/boundary/CMakeFiles/boundary.dir/periodicCC.cc.o
Linking CXX static library libwriter.a
[ 54%] Linking CXX static library libtseries.a
Built target writer
[ 56%] Building CXX object lib/force/CMakeFiles/force.dir/rotatingConv.cc.o
[ 56%] Built target tseries
[ 59%] Building CXX object lib/field/CMakeFiles/field.dir/plainsf.cc.o
Linking CXX static library libparser.a
[ 59%] Built target parser
[ 61%] Building CXX object lib/force/CMakeFiles/force.dir/randomForcing.cc.o
[ 63%] Building CXX object lib/initial/CMakeFiles/initial.dir/channelRand.cc.o
[ 65%] Building CXX object lib/boundary/CMakeFiles/boundary.dir/periodicFC.cc.o
[ 68%] Building CXX object lib/field/CMakeFiles/field.dir/plainvf.cc.o
Linking CXX static library libgrid.a
[ 68%] Built target grid
[ 70%] Building CXX object lib/boundary/CMakeFiles/boundary.dir/neumannCC.cc.o
[ 72%] Building CXX object lib/field/CMakeFiles/field.dir/derivative.cc.o
[ 75%] Building CXX object lib/boundary/CMakeFiles/boundary.dir/neumannFC.cc.o
[ 77%] Building CXX object lib/boundary/CMakeFiles/boundary.dir/hotPlateCC.cc.o
Linking CXX static library libinitial.a
[ 77%] Built target initial
In file included from /Users/odoronin/saras-master/lib/boundary/neumannCC.cc:43:
/Users/odoronin/saras-master/lib/boundary/boundary.h:183:20: warning: private field 'fieldValue' is not used [-Wunused-private-field]
        const real fieldValue;
                   ^
In file included from /Users/odoronin/saras-master/lib/boundary/neumannFC.cc:43:
/Users/odoronin/saras-master/lib/boundary/boundary.h:200:20: warning: private field 'fieldValue' is not used [-Wunused-private-field]
        const real fieldValue;
                   ^
In file included from /Users/odoronin/saras-master/lib/field/plainvf.cc:43:
/Users/odoronin/saras-master/lib/field/plainvf.h:51:21: warning: private field 'gridData' is not used [-Wunused-private-field]
        const grid &gridData;
                    ^
Linking CXX static library libforce.a
[ 77%] Built target force
1 warning generated.
1 warning generated.
Linking CXX static library libpoisson.a
[ 77%] Built target poisson
1 warning generated.
Linking CXX static library libfield.a
Linking CXX static library libboundary.a
[ 77%] Built target field
Scanning dependencies of target saras_test
[ 77%] Built target boundary
[ 79%] [ 81%] [ 84%] [ 86%] [ 88%] [ 90%] [ 93%] [ 95%] Building CXX object src/tests/CMakeFiles/saras_test.dir/main.cc.o
Building CXX object src/tests/CMakeFiles/saras_test.dir/global/unittest.cc.o
[ 97%] Building CXX object src/tests/CMakeFiles/saras_test.dir/__/solvers/hydro/hydro.cc.o
[100%] Building CXX object src/tests/CMakeFiles/saras_test.dir/__/solvers/hydro/hydro_d2.cc.o
Building CXX object src/tests/CMakeFiles/saras_test.dir/units/field_test.cc.o
Building CXX object src/tests/CMakeFiles/saras_test.dir/__/solvers/hydro/hydro_d3.cc.o
Building CXX object src/tests/CMakeFiles/saras_test.dir/units/differ_test.cc.o
Building CXX object src/tests/CMakeFiles/saras_test.dir/units/nlin_test.cc.o
Building CXX object src/tests/CMakeFiles/saras_test.dir/units/poisson_test.cc.o
Building CXX object src/tests/CMakeFiles/saras_test.dir/units/hydro_test.cc.o
In file included from /Users/odoronin/saras-master/src/tests/global/unittest.cc:43:
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:29: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                            ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:29: error: use of undeclared identifier 'real'
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:54: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                                                     ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:80: error: unknown type name 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                                                                               ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:53:18: error: unknown type name 'real'
void printResult(real computedValue, real errorTolerance);
                 ^
In file included from /Users/odoronin/saras-master/src/tests/units/poisson_test.cc:44:
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:29: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                            ^
In file included from /Users/odoronin/saras-master/src/tests/main.cc:43:
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:29: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                            ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:53:38: error: unknown type name 'real'
void printResult(real computedValue, real errorTolerance);
                                     ^
In file included from /Users/odoronin/saras-master/src/tests/units/nlin_test.cc:44:
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:29: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                            ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:29: error: use of undeclared identifier 'real'
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:29: error: use of undeclared identifier 'real'
In file included from /Users/odoronin/saras-master/src/tests/units/differ_test.cc:44:
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:29: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                            ^
/Users/odoronin/saras-master/src/tests/global/unittest.cc:45:29: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol) {
                            ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:29: error: use of undeclared identifier 'real'
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:54: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                                                     ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:29: error: use of undeclared identifier 'real'
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:54: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                                                     ^
/Users/odoronin/saras-master/src/tests/global/unittest.cc:45:29: error: use of undeclared identifier 'real'
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:80: error: unknown type name 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                                                                               ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:80: error: unknown type name 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                                                                               ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:54: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                                                     ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:53:18: error: unknown type name 'real'
void printResult(real computedValue, real errorTolerance);
                 ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:54: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                                                     ^
/Users/odoronin/saras-master/src/tests/global/unittest.cc:45:54: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol) {
                                                     ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:53:18: error: unknown type name 'real'
void printResult(real computedValue, real errorTolerance);
                 ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:53:38: error: unknown type name 'real'
void printResult(real computedValue, real errorTolerance);
                                     ^
In file included from /Users/odoronin/saras-master/src/tests/units/field_test.cc:44:
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:29: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                            ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:80: error: unknown type name 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                                                                               ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:80: error: unknown type name 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                                                                               ^
/Users/odoronin/saras-master/src/tests/global/unittest.cc:45:80: error: unknown type name 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol) {
                                                                               ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:53:38: error: unknown type name 'real'
void printResult(real computedValue, real errorTolerance);
                                     ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:53:18: error: unknown type name 'real'
void printResult(real computedValue, real errorTolerance);
                 ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:53:18: error: unknown type name 'real'
void printResult(real computedValue, real errorTolerance);
                 ^
/Users/odoronin/saras-master/src/tests/global/unittest.cc:47:5: error: unknown type name 'real'
    real errorVal;
    ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:29: error: use of undeclared identifier 'real'
/Users/odoronin/saras-master/src/tests/global/unittest.h:53:38: error: unknown type name 'real'
void printResult(real computedValue, real errorTolerance);
                                     ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:53:38: error: unknown type name 'real'
void printResult(real computedValue, real errorTolerance);
                                     ^
/Users/odoronin/saras-master/src/tests/global/unittest.cc:81:18: error: unknown type name 'real'
void printResult(real computedValue, real errorTolerance) {
                 ^
/Users/odoronin/saras-master/src/tests/global/unittest.cc:81:38: error: unknown type name 'real'
void printResult(real computedValue, real errorTolerance) {
                                     ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:54: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                                                     ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:80: error: unknown type name 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                                                                               ^
In file included from /Users/odoronin/saras-master/src/tests/units/hydro_test.cc:44:
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:29: error: use of undeclared identifier 'real'
/Users/odoronin/saras-master/src/tests/global/unittest.h:53:18:void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                            ^error
: unknown type name 'real'
void printResult(real computedValue, real errorTolerance);
                 ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:53:38: error: unknown type name 'real'
void printResult(real computedValue, real errorTolerance);
                                     ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:29: error: use of undeclared identifier 'real'
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:54: error: use of undeclared identifier 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                                                     ^
/Users/odoronin/saras-master/src/tests/global/unittest.h:51:80: error: unknown type name 'real'
void testError(blitz::Array<real, 3> A, blitz::Array<real, 3> B, int errorMom, real errorTol);
                                                                               ^
[…]
