### 2021-03-08 #2095 @roshansamuel
Hi @olgadoronina, the error you are getting is most likely due to MPI failing to initialize at ``MPI_Init()``. I suspect that not just SARAS, but any MPI code might give the same error. Can you please confirm if a simple MPI example code runs on your system? If you are running MacOS, [this link](https://stackoverflow.com/questions/23112515/mpich2-gethostbyname-failed) might help.

The second set of errors is from enabling the TEST_RUN flag in the build script. Kindly disable the flag as it was used only during an early stage of development and is no longer in use. We will remove the flag to avoid confusion in the future. To test SARAS, please run ``testSaras.sh`` script in the ``tests/`` folder. You can also test the multi-grid solver by running ``testPoisson.sh`` script inside the same folder.
