### 2021-03-08 #2095 @nickwimer
Hi @roshansamuel I have had no more success with installing SARAS on MacOS. I have also attempted install on Eagle supercomputer and again, it failed.

Finally, I switched to NERSC Cori machine and everything installed correctly as described in your documentation. I have tried running the `testSaras.sh` script and it appears to run correctly, however when I try to run `validate_ldc.py`, the error is: "Could not open file output/Soln_0030.0000.h5"

`testPoisson.sh` runs without any errors.

I should also note that `scipy` is needed to run `validate_ldc.py` though it is not listed in the README.
