+++ tests/ldcTest/validate_ldc.py
@@ -217,7 +217,7 @@ if __name__ == "__main__":
 
     parseYAML("input/parameters.yaml")
 
-    loadData(30.0)
+    loadData(50.0)
 
     loadGhia()
 
