### 2021-03-09 #2095 @roshansamuel
Hi @nickwimer the error in opening ``Soln_0030.0000.h5`` was due to a bug in ``validate_ldc.py`` and it has been corrected. The README has also been updated as you suggested. We thank you for inputs.

It would be very helpful if you can post the error you faced when installing on Eagle supercomputer and MacOS. The error in ``MPI_Init()`` in MacOS can be resolved using the suggestion in [this link](https://stackoverflow.com/questions/23112515/mpich2-gethostbyname-failed). We have tried to address as many installation issues as possible, and would like to eliminate any outstanding issues.
