### 2021-05-19 bd1d196
Updated CMakeLists and parser to handle both old and new versions of yaml-cpp
