+++ CMakeLists.txt
@@ -62,6 +62,19 @@ message (STATUS "The parent directory is ${PARENT_DIR}")
 message (STATUS "Install folder is ${INSTALL_DIR}")
 message (STATUS "All headers are in ${HEADER_DIR}")
 
+# Search for yaml-cpp library
+find_package(yaml-cpp)
+if (yaml-cpp_FOUND)
+    if (yaml-cpp_VERSION GREATER 0.3)
+        message (STATUS "Found YAML Cpp ${yaml-cpp_VERSION}")
+    else ()
+        message (STATUS "Found YAML Cpp ${yaml-cpp_VERSION}. Using deprecated YAML parsing commands.")
+        add_definitions(-DYAML_LEGACY)
+    endif ()
+else ()
+    message (FATAL_ERROR "Did not find YAML-Cpp library")
+endif ()
+
 # Check if user defined flags are set
 if (NOT DEFINED PLANAR)
     message (STATUS "PLANAR flag not set. Compiling for default 3D run")
@@ -130,3 +143,7 @@ unset (REAL_SINGLE CACHE)
 
 # Remove the TEST_POISSON variable from cache to force user to manually set the flag for testing Poisson solver
 unset (TEST_POISSON CACHE)
+
+# Remove variables associated with finding yaml-cpp version from cache
+unset (yaml-cpp_FOUND CACHE)
+unset (yaml-cpp_VERSION CACHE)
