### 2021-06-07 45741b2
Made build process more transparent. Removed compile-script. Updated README
