+++ compile/compileSaras.sh
@@ -1,146 +1,0 @@
-#!/bin/bash
-
-###################################################################################
- # Saras
- # 
- # Copyright (C) 2019, Mahendra K. Verma
- #
- # All rights reserved.
- # 
- # Redistribution and use in source and binary forms, with or without
- # modification, are permitted provided that the following conditions are met:
- #     1. Redistributions of source code must retain the above copyright
- #        notice, this list of conditions and the following disclaimer.
- #     2. Redistributions in binary form must reproduce the above copyright
- #        notice, this list of conditions and the following disclaimer in the
- #        documentation and/or other materials provided with the distribution.
- #     3. Neither the name of the copyright holder nor the
- #        names of its contributors may be used to endorse or promote products
- #        derived from this software without specific prior written permission.
- # 
- # THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- # ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- # WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- # DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
- # ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- # (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- # LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- # ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- # SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- #
- ##################################################################################
- ##
- ##! \file compileSaras.sh
- #
- #   \brief Shell script to automatically compile and run SARAS
- #
- #   \author Roshan Samuel
- #   \date Jan 2020
- #   \copyright New BSD License
- #
- ##################################################################################
- ##
-
-
-############## USER SET PARAMETERS - COMMENT/UNCOMMENT AS NECESSARY ###############
-
-# If below flag is uncommented, the script automatically runs the executable after
-# compilation by issuing the mpirun command
-#EXECUTE_AFTER_COMPILE="EXECUTE"
-
-# Number of processors to be used when running SARAS. This parameter is used only
-# if the EXECUTE_AFTER_COMPILE parameter is enabled.
-PROC=4
-
-# This parameter has to be enabled to use the SARAS executable for 2D simulations.
-#PLANAR="PLANAR"
-
-# SARAS supports computations with either double or single precision floating point
-# values. This parameter must be either SINGLE or DOUBLE
-REAL_TYPE="DOUBLE"
-
[…]
