+++ README.md
@@ -8,59 +8,61 @@ All the source and library files for the Saras solver are contained in the follo
 
 * ``./src/`` - contains all the different solvers available in Saras
 * ``./lib/`` - contains all the libraries used by the solvers in ``./src/`` folder
-* ``./compile/`` - contains the installation script to build and compile the solver
-* ``./output/`` - the solution files written by the solver are contained in this folder, it also contains Python post-processing scripts
+* ``./input/`` - contains the parameters to be read by the solver
+* ``./output/`` - the solution files are written into this folder, it also contains Python post-processing scripts
 
 ## Installing SARAS
 
 ``SARAS`` relies on a few libraries for its calculations.
 Therefore the first step towards building ``SARAS`` is to install the following dependencies:
 
-* ``blitz`` - All array manipulations are performed using the Blitz++ library
 * ``cmake`` - Necessary to build the ``saras`` executable
 * ``mpich`` - For parallel computation using MPI
-* ``yaml`` - The input parameters are stored in a YAML file which needs to be parsed using yaml-cpp library.
 * ``hdf5`` - The output files are written in HDF5 format
+* ``blitz`` - All array manipulations are performed using the Blitz++ library
+* ``yaml-cpp`` - The input parameters are stored in a YAML file which needs to be parsed using yaml-cpp library.
 
-Packages like ``cmake``, ``mpich`` and ``hdf5`` can be installed from the OS package manager,
-while ``yaml`` and ``blitz`` can be downloaded and installed manually.
+Packages like ``cmake``, ``mpich``, ``yaml-cpp``, and ``hdf5`` can be installed from the OS package manager.
+However, the ``blitz`` package will have to be downloaded and installed manually.
 
-However, we provide below an alternative series of steps that will install all the libraries in the user's home directory.
-This method has three advantages:
-
-* It does not require administrator (``sudo``) privileges,
-* It will not disturb pre-existing packages already installed on the system
-* The method has been tested on both Linux (Ubuntu 14.04 and above), and MacOS (Mojave)
+If you do not have the administrator privileges required to install packages using the OS package manager,
+you can also install them in your home folder.
+This also offers the potential advantage of not disturbing pre-existing packages already installed on the system.
+The steps listed below explain this method of installation.
+That said, libraries like ``cmake``, ``MPICH`` and ``HDF5`` are normally available on most computing systems.
+Please check if these packages are already installed, and if they are, you can skip their installation steps.
 
 ### Download all the dependencies
 
-All the required packages can be downloaded from our [lab website](http://turbulencehub.org), and installed over a terminal.
+All the required packages can be downloaded and installed over a terminal.
 It is advisable to create a temporary directory where the packages can be downloaded and extracted.
-After navigating to the temporary directory, download the packages using ``wget`` on Linux, or ``curl`` on MacOS.
-For instance, on MacOS, the following lines will download ``CMake``, ``Blitz++``, ``yaml-cpp``, ``MPICH`` and ``HDF5`` packages respectively:
+After navigating to the temporary directory, download the packages using ``wget`` command on Linux, or ``curl`` command on MacOS.
+``CMake``, ``MPICH`` and ``HDF5`` packages can be downloaded from their respective sites, and extracted by the ``tar`` command.
 
 ```
[…]

[…]
