### 2021-06-07 3cf3f90
Updated README with minor corrections
