+++ README.md
@@ -1,12 +1,12 @@
 # Saras - Finite difference solver
 
-Saras is an MPI parallelized Navier-Stokes equation solver written in C++.
+SARAS is an MPI parallelized Navier-Stokes equation solver written in C++.
 It uses the finite-difference method for calculating spatial derivatives and parallelized geometric multi-grid method for solving
 the pressure Poisson equation.
 
-All the source and library files for the Saras solver are contained in the following directories:
+All the relevant files of the solver are contained in the following directories:
 
-* ``./src/`` - contains all the different solvers available in Saras
+* ``./src/`` - contains the different solvers available in SARAS
 * ``./lib/`` - contains all the libraries used by the solvers in ``./src/`` folder
 * ``./input/`` - contains the parameters to be read by the solver
 * ``./output/`` - the solution files are written into this folder, it also contains Python post-processing scripts
@@ -26,9 +26,10 @@ Packages like ``cmake``, ``mpich``, ``yaml-cpp``, and ``hdf5`` can be installed 
 However, the ``blitz`` package will have to be downloaded and installed manually.
 
 If you do not have the administrator privileges required to install packages using the OS package manager,
-you can also install them in your home folder.
+you can install them in your home folder instead.
 This also offers the potential advantage of not disturbing pre-existing packages already installed on the system.
 The steps listed below explain this method of installation.
+
 That said, libraries like ``cmake``, ``MPICH`` and ``HDF5`` are normally available on most computing systems.
 Please check if these packages are already installed, and if they are, you can skip their installation steps.
 
@@ -37,7 +38,10 @@ Please check if these packages are already installed, and if they are, you can s
 All the required packages can be downloaded and installed over a terminal.
 It is advisable to create a temporary directory where the packages can be downloaded and extracted.
 After navigating to the temporary directory, download the packages using ``wget`` command on Linux, or ``curl`` command on MacOS.
-``CMake``, ``MPICH`` and ``HDF5`` packages can be downloaded from their respective sites, and extracted by the ``tar`` command.
+[CMake](https://cmake.org/download/),
+[MPICH](https://www.mpich.org/downloads/) and
+[HDF5](https://portal.hdfgroup.org/display/support/HDF5+1.8.21)
+packages can be downloaded from their respective sites, and extracted by the ``tar`` command.
 
 ```
 wget https://github.com/Kitware/CMake/releases/download/v3.20.3/cmake-3.20.3.tar.gz
@@ -64,7 +68,7 @@ the packages can be downloaded using the links listed above, and transferred ove
 > the $LANG environment variable has to be set.
 > For this you need to execute the command: `export LANG=en_US.UTF-8`
 
-Once all the downloaded packages have been extracted, create an install location in the user's home folder if it doesn't exist already.
+Once all the downloaded packages have been extracted, create an install location in your home folder if it doesn't exist already.
 Usually, packages are installed in ``$HOME/local/`` directory.
 The next steps will assume that the folder ``local/`` exists in the user's home directory.
 
@@ -90,7 +94,6 @@ Accordingly, set the following environment variables:
 ```
 export PATH=$HOME/local/bin:$PATH
 export PKG_CONFIG_PATH=$HOME/local/lib/pkgconfig:$PKG_CONFIG_PATH
-export HDF5_ROOT=$HOME/local
 export CPATH=$HOME/local/include/:$CPATH
 export LD_LIBRARY_PATH=$HOME/local/lib:$LD_LIBRARY_PATH
 export LIBRARY_PATH=$HOME/local/lib:$LIBRARY_PATH
@@ -105,7 +108,7 @@ To permanently add the path variables, the above lines may be appended to the sh
 
[…]
