### 2021-06-07 #2095 @roshansamuel
Hi @olgadoronina, SARAS is now compatible with the latest versions of all the dependent libraries. You do not have to install the libraries from turbulencehub now. The README has been updated to reflect this. All the download links now point to the latest tarballs of each library. Ideally, it is best to use the libraries supplied by the package manager of your OS (for CMake, MPICH, HDF5 etc.).

The Blitz and yaml-cpp libraries can now be downloaded from their respective Git repositories. The installation instructions have been updated for this as well.

The process of building SARAS has also been updated. You can make a build directory and run CMake from within it now. The compileScript, which seemed to be a source of confusion, has been removed. Hope this makes the installation of SARAS more aligned with that of other standard packages.
