+++ lib/parallel/mpidata.cc
@@ -94,27 +94,22 @@ void mpidata::createSubarrays(const blitz::TinyVector<int, 3> globSize,
 
     globCopy = globSize;
 
-    // CREATING SUBARRAYS FOR TRANSFER ACROSS THE 4 FACES OF EACH SUB-DOMAIN
-
     /************************************************************** NOTE ************************************************************\
      * MPI Subarrays assume that the starting index of arrays is 0, 0, 0                                                            *
      * But the arrays used here through Blitz start with the index (-padX, -padY, -padZ)                                            *
      * Hence the saStarts variable is shifted accordingly                                                                           *
     \********************************************************************************************************************************/
 
+    // CREATING SUBARRAYS FOR TRANSFER ACROSS THE 4 FACES OF EACH SUB-DOMAIN
+
     //*****************************************************! ALONG XI-DIRECTION !***************************************************//
     // SEND SUB-ARRAY ON LEFT SIDE
     saStarts = padWidth;
     loclSize = coreSize;            loclSize(0) = padWidth(0);
 
-    // STAGGERED GRID SHARE A POINT ACROSS SUB-DOMAIN BOUNDARIES AND HENCE SENDS A SLIGHTLY DIFFERENT DATA-SET
-    // HOWEVER, IF THE DOMAIN IS PERIODIC (DEFAULT), THE DATA TO BE SENT ON THE LEFT IS FROM THE WALL POINT ITSELF
-    // WHEN USING Method 3 OF setBulkSlice FUNCTION IN field.cc, THE BELOW MODIFICATION APPLIES TO ALL RANKS
-    // ELSE, IT APPLIES ONLY TO xRank > 0
-    //if (xStag and rankData.xRank > 0) {
-    if (xStag) {
-        saStarts(0) += padWidth(0);
-    }
+    // STAGGERED GRID SHARE A POINT ACROSS SUB-DOMAIN BOUNDARIES,
+    // AND HENCE SENDS A SLIGHTLY DIFFERENT DATA-SET
+    if (xStag) saStarts(0) += 1;
 
     MPI_Type_create_subarray(3, globCopy.data(), loclSize.data(), saStarts.data(), MPI_ORDER_C, MPI_FP_REAL, &sendSubarrayX0);
     MPI_Type_commit(&sendSubarrayX0);
@@ -126,15 +121,13 @@ void mpidata::createSubarrays(const blitz::TinyVector<int, 3> globSize,
     MPI_Type_create_subarray(3, globCopy.data(), loclSize.data(), saStarts.data(), MPI_ORDER_C, MPI_FP_REAL, &recvSubarrayX0);
     MPI_Type_commit(&recvSubarrayX0);
 
-
     // SEND SUB-ARRAY ON RIGHT SIDE
     saStarts = padWidth;            saStarts(0) = coreSize(0);
     loclSize = coreSize;            loclSize(0) = padWidth(0);
 
-    // STAGGERED GRID SHARE A POINT ACROSS SUB-DOMAIN BOUNDARIES AND HENCE SENDS A SLIGHTLY DIFFERENT DATA-SET
-    if (xStag) {
[…]
