+++ tests/testPoisson.sh
@@ -54,10 +54,10 @@ fi
 cd build
 
 # Run cmake with necessary flags for 2D Poisson test
-#CC=mpicc CXX=mpicxx cmake ../../ -DPLANAR=ON -DTEST_POISSON=ON -DREAL_DOUBLE=ON
+#CC=mpicc CXX=mpicxx cmake ../../ -DPLANAR=ON -DTEST_POISSON=ON
 
 # Run cmake with necessary flags for 3D Poisson test
-CC=mpicc CXX=mpicxx cmake ../../ -DTEST_POISSON=ON -DREAL_DOUBLE=ON
+CC=mpicc CXX=mpicxx cmake ../../ -DTEST_POISSON=ON
 
 # Compile
 make -j8
