### 2021-06-14 e006574
Fixed bugs and updated documentation in derivative class
