+++ lib/field/derivative.h
@@ -33,7 +33,7 @@
  *
  *  \brief Class declaration of derivative
  *
- *  \author Ali Asad, Roshan Samuel
+ *  \author Roshan Samuel, Ali Asad
  *  \date Nov 2019
  *  \copyright New BSD License
  *
@@ -46,7 +46,6 @@
 #include <blitz/array.h>
 #include <blitz/array/stencil-et.h>
 #include <blitz/array/stencilops.h>
-#include <string>
 
 #include "field.h"
 #include "grid.h"
@@ -57,30 +56,32 @@ class derivative {
 
         const field &F;
 
-        real invDelx, invDely, invDelz;
+        real ihx, ihy, ihz;
+        real ihx2, ihy2, ihz2;
 
         blitz::firstIndex i;
         blitz::secondIndex j;
         blitz::thirdIndex k;    
 
         blitz::Range fullRange;
+        blitz::Range xRange, yRange, zRange;
 
         blitz::Array<real, 1> x_Metric, y_Metric, z_Metric;
         blitz::Array<real, 1> xxMetric, yyMetric, zzMetric;
         blitz::Array<real, 1> x2Metric, y2Metric, z2Metric;
 
-        blitz::Array<real, 3> tempMat;
+        blitz::Array<real, 3> tmpArray;
 
     public:
         derivative(const grid &gridData, const field &F);
 
-        void calcDerivative1_x(blitz::Array<real, 3> outputMat);
-        void calcDerivative1_y(blitz::Array<real, 3> outputMat);
-        void calcDerivative1_z(blitz::Array<real, 3> outputMat);
+        void calcDerivative1_x(blitz::Array<real, 3> outArray);
+        void calcDerivative1_y(blitz::Array<real, 3> outArray);
+        void calcDerivative1_z(blitz::Array<real, 3> outArray);
 
-        void calcDerivative2xx(blitz::Array<real, 3> outputMat);
-        void calcDerivative2yy(blitz::Array<real, 3> outputMat);
-        void calcDerivative2zz(blitz::Array<real, 3> outputMat);
+        void calcDerivative2xx(blitz::Array<real, 3> outArray);
+        void calcDerivative2yy(blitz::Array<real, 3> outArray);
+        void calcDerivative2zz(blitz::Array<real, 3> outArray);
 };
 
 /**
@@ -88,8 +89,11 @@ class derivative {
  *  \class derivative derivative.h "lib/derivative.h"
  *  \brief Derivative class to perform finite difference operations on the data stored in field
  *
- *  It contains functions to perform the finite difference operations with constant grid spacing.
- *  For many classes of this solver, empty destructors are removed. Refer reference [5] of README for more details.
+ *  It contains functions to perform the finite difference operations on fields.
+ *  For fields on non-uniform grids, the derivatives are transformed with appropriate
+ *  grid derivatives taken from the grid class.
+ *  For many classes of SARAS, empty destructors were removed.
+ *  Refer reference [3] of General Articles in README for more details.
  ********************************************************************************************************************************************
  */
 
