### 2021-06-14 7327b7f
Updated README with instructions to handle OpenSSL error while installing CMake
