### 2021-06-14 b31dbbb
Minor formatting updates to README
