### 2021-08-15 #2095 @roshansamuel
Hi @kyleniemeyer, can something be done to finish the review? We have installed and run SARAS on various workstations and clusters without encountering hiccups.

Blitz++ and yaml-cpp are the only dependencies that actually require manual installation to run SARAS. The other libraries (CMake, MPICH, and HDF5) are very commonly available in most software repositories and don't usually require manual installation.

Previously there was a restriction on version compatibility of yaml-cpp. We subsequently addressed this issue and made SARAS compatible with the latest libraries.

Two publications of results using SARAS are out, and a third one is in preparation. We are unable to cite SARAS appropriately in all these works. It has now been 1.5 years since the submission of our software to JoSS. We understand that the pandemic has placed significant impediments in the review process. However, we would be highly obliged if this review can be completed after so long a time period.
