### 2021-08-16 #2095 @kyleniemeyer
@roshansamuel yes, I think we can move forward with this—thanks for your patience! I'm going to take a final look at your paper.

In the meantime, can you please archive the software repo (e.g., on Zenodo or Figshare) and provide the DOI here?
