### 2021-08-16 #2095 @roshansamuel
@kyleniemeyer thank you very much.

A tagged release of SARAS has been archived on Zenodo and its DOI is [https://doi.org/10.5281/zenodo.5205320](https://doi.org/10.5281/zenodo.5205320)
