### 2021-08-16 #2095 @kyleniemeyer
Hi @roshansamuel, just a few comments about the article:

1. Can you please change the "Summary" section to "Summary and Statement of Need", or otherwise break out part of the section as an explicit "Statement of Need" section? We have made that a required element of JOSS papers.
2. Regarding the sentence "We are planning to conduct a scaling analysis of SARAS and compare its performance with TARANG in the near future": is this still true? If this has been done in the time since original submission, it would be good to add this result and citation, or modify the language here if not.
3. In the Ghia et al. citation, I think you need brackets around Re in the article title, to preseve capitalization (i.e., `{Re}`)
