### 2021-08-16 bb3fd54
Updated paper with statement of need, minor corrections
