+++ resources/paper.bib
@@ -63,7 +63,7 @@
 }
 
 @article{Ghia:JCP1982,
-    title={High-Re solutions for incompressible flow using the Navier-Stokes equations and a multigrid method},
+    title={High-{Re} solutions for incompressible flow using the Navier-Stokes equations and a multigrid method},
     author={Ghia, U. and Ghia, K.N. and Shin, C.T.},
     journal={J. Comput. Phys.},
     volume={48},
