+++ CMakeLists.txt
@@ -89,7 +89,7 @@ if (yaml-cpp_FOUND)
 else ()
     find_library(YAML_FOUND yaml-cpp libyaml-cpp)
     if (YAML_FOUND)
-        message (STATUS "Found YAML Cpp with find_library(). You may have to configure with DYAML_LEGACY")
+        message (STATUS "Found YAML Cpp but could not determine version. You may have to configure with DYAML_LEGACY")
     else ()
         message (WARNING "Could not determine YAML-Cpp version. You may have to configure with DYAML_LEGACY")
     endif ()
