### 2021-08-16 #2095 @roshansamuel
Hi @kyleniemeyer, thank you for your corrections. We have updated the paper as you said.

1. The "Summary" section has been broken up to add a "Statement of Need" section.
2. We had performed a preliminary scaling analysis using 1024 cores, but we need to optimize the solver further before moving on to higher numbers of cores. We now mention this in the paper.
3. The bib file for the Ghia citation has been corrected. Thank you for pointing this out.
